Adaptive_bandwidth

Computes the data-driven bandwidth using a method based on the spectral density operator which adapts to the functional data.

Usage

adaptive_bandwidth(
  X,
  kernel = bartlett_kernel,
  name = NULL,
  order = NULL,
  weighting = NULL
)

Arguments

X

A dfts object or data which can be automatically converted to that format. See dfts().

kernel

Kernel function. No additional parameters are needed for bartlett_kernel(), parzen_kernel(), tukey_hanning_kernel(), and quadratic_spectral_kernel().

name, order, weighting

Additional parameters if non-standard kernels, e.g. those not in fChange, are used. See references for the definitions. Name is extracted from the kernel name to select order/weighting when not given, if the function aligns with the recommended functions, see kernel parameter.

Value

Scalar value of the data-adapted bandwidth.

References

Rice, G., & Shang, H. L. (2017). A Plug‐in Bandwidth Selection Procedure for Long‐Run Covariance Estimation with Stationary Functional Time Series. Journal of Time Series Analysis, 38(4), 591–609.

See also

Examples

adaptive_bandwidth(generate_brownian_motion(100))
#> [1] 4.175179
adaptive_bandwidth(electricity, parzen_kernel)
#> [1] 39.99534