dfts Objects

Define data as, check data if, and convert data to dfts object. The data may be from another popular implementation.

Usage

dfts(X, name = NULL, labels = NULL, intratime = NULL, inc.warnings = TRUE)

as.dfts(X)

is.dfts(X)

Arguments

X

Data to convert into dfts object. Options include: data.frame, matrix, array, fda::fd, fda.usc::fdata, rainbow::fts (used in ftsa), rainbow::fds (used in ftsa), funData::funData, and fChange::dfts. For a matrix, each column is a unique observation, at the rows are the observed intra-observation (i.e. resolution) points.

name

String for the name of the object. Defaults to the name of the input variable.

labels

Labels for the observations. Defaults to the column names or names inside of the object X.

intratime

Vector of numerics indicating the points for each observation. Defaults to even spacing on [0,1], or those included in the object. These may be unevenly spaced.

inc.warnings

Boolean on if warnings should be given. Defaults to TRUE,

Value

dfts / as.dfts: dfts object

is.dfts: Boolean indicating if x is a dfts object or not

Examples

bm <- dfts(generate_brownian_motion(100, c(0,0.1,0.25,0.5,1)))

# Electricity
result <- dfts(electricity)
result <- as.dfts(electricity)
result <- is.dfts(electricity)