Difference dfts

Difference the functional data at some lag / to some degree.

Usage

# S3 method for class 'dfts'
diff(x, lag = 1L, differences = 1L, ...)

Arguments

x

A dfts object or data which can be automatically converted to that format. See dfts().

lag

An integer indicating which lag to use.

differences

An integer indicating the order of the difference.

...

Further arguments to be passed to or from methods.

Value

dfts object with differenced values

Examples

result <- diff(electricity, lag=1)
result1 <- diff(electricity, differences=2)