Generate Data via Ornstein-Uhlenbeck Process

Generate autoregressive process with errors according the Ornstein-Uhlenbeck process.

Usage

generate_ornstein_uhlenbeck(N, v, rho = 0)

Arguments

N

Numeric for the number of observations.

v

Numeric for resolution of data or a vector specifying the observation points.

rho

Numeric which indicates the dependence on the previous curve.

Value

A dfts object for the generated data.

Examples

generate_ornstein_uhlenbeck(N=100,v=20)
#> data 
#> dimension: 20 x 100
#> season: 1
#> fparam (20): 0 0.05263158 0.1052632 0.1578947 0.2105263 0.2631579 ... 
#> labels (100): X1 X2 X3 X4 X5 X6 ... 
#> 
#> Example data
#>            [,1]      [,2]      [,3]        [,4]      [,5]      [,6]
#> [1,] 0.25708742 0.8698224 1.1235103 -0.42382498 1.1183894 0.1337938
#> [2,] 0.07877327 0.9631859 1.1746122 -0.35317075 1.1969059 0.4075599
#> [3,] 0.29009782 0.9243200 1.0623083 -0.12229039 1.0583682 0.2748367
#> [4,] 0.26328019 1.1340691 1.1054126 -0.01808502 0.9604892 0.6374935
#> [5,] 0.15478423 1.0922602 1.0893714 -0.34480356 0.9856522 0.7124355
#> [6,] 0.26975312 0.8817843 0.5798364  0.11500793 1.0670967 0.9053514