Imputation

Impute missing values in functional data.

Usage

impute(
  X,
  method = c("zero", "mean_obs", "median_obs", "mean_data", "median_data", "linear",
    "functional"),
  obs_share_data = FALSE
)

Arguments

X

A dfts object or data which can be automatically converted to that format. See dfts().

method

String to indicate method of imputation.

zero: Fill missing with 0

mean_obs: Fill missing with mean of each observation

median_obs: Fill missing with median of each observation

mean_data: Fill missing with mean at time through all data

median_data: Fill missing with median at time through all data

linear: Linearly interpolate missing data

obs_share_data

Boolean in linear interpolation that indicates if data should be shared across observations. For example, is the end of observation i related to the start of observation i+1. Default is FALSE, which suggests independence. If true, the distance between the end and start of observations is taken to be the mean of the intratime.

Value

dfts object with interpolated missing data

Examples

temp <- data.frame(c(NA,NA,3:9,NA),
                   c(NA,stats::rnorm(2),NA, stats::rnorm(6)),
                   stats::rnorm(10),
                   c(stats::rnorm(4),rep(NA,3), stats::rnorm(3)),
                   rep(NA,10),
                   c(stats::rnorm(1), rep(NA,9)),
                   c(stats::rnorm(9),NA),
                   stats::rnorm(10),
                   stats::rnorm(10),
                   c(NA,NA,3:9, NA))
impute(temp, method='mean_obs')
#> $data
#>       [,1]         [,2]        [,3]       [,4] [,5]       [,6]       [,7]
#>  [1,]    6  0.307985975  3.51781068 -0.4733455   NA 0.07974821 -1.2884072
#>  [2,]    6  0.324331233  1.15199039  0.4501548   NA 0.07974821  0.4744997
#>  [3,]    3  0.628674056  0.38842370  0.5037310   NA 0.07974821  1.1574781
#>  [4,]    4  0.307985975 -0.32097173 -0.1552386   NA 0.07974821  0.1428439
#>  [5,]    5  0.301473839 -2.39135311 -0.2271112   NA 0.07974821  0.1008368
#>  [6,]    6  1.790640499  0.89829813 -0.2271112   NA 0.07974821  0.8609186
#>  [7,]    7 -0.005856082 -0.11913019 -0.2271112   NA 0.07974821  2.5377739
#>  [8,]    8  0.530457470 -1.81122400 -0.4561464   NA 0.07974821  0.4499884
#>  [9,]    9 -1.112591695  0.35752554 -1.0144646   NA 0.07974821  1.9895189
#> [10,]    6  0.006758480 -0.09005252 -0.4444690   NA 0.07974821  0.7139390
#>             [,8]       [,9] [,10]
#>  [1,]  0.4406636  1.1778013     6
#>  [2,] -1.0379310  1.9360774     6
#>  [3,]  0.1665210 -0.4595392     3
#>  [4,] -0.6092184 -0.1425464     4
#>  [5,]  0.5813857  0.4308495     5
#>  [6,] -0.2896958 -2.0217243     6
#>  [7,] -1.4210328  0.4704179     7
#>  [8,] -0.6940940  0.3414290     8
#>  [9,] -0.4123855 -1.0954659     9
#> [10,] -0.2301435  1.4910926     6
#> 
#> $name
#> [1] "X"
#> 
#> $labels
#>  [1] "c.NA..NA..3.9..NA."                             
#>  [2] "c.NA..stats..rnorm.2...NA..stats..rnorm.6.."    
#>  [3] "stats..rnorm.10."                               
#>  [4] "c.stats..rnorm.4...rep.NA..3...stats..rnorm.3.."
#>  [5] "rep.NA..10."                                    
#>  [6] "c.stats..rnorm.1...rep.NA..9.."                 
#>  [7] "c.stats..rnorm.9...NA."                         
#>  [8] "stats..rnorm.10..1"                             
#>  [9] "stats..rnorm.10..2"                             
#> [10] "c.NA..NA..3.9..NA..1"                           
#> 
#> $intratime
#>  [1] 0.0000000 0.1111111 0.2222222 0.3333333 0.4444444 0.5555556 0.6666667
#>  [8] 0.7777778 0.8888889 1.0000000
#> 
#> attr(,"class")
#> [1] "dfts"
impute(temp, method='linear', obs_share_data=TRUE)
#> $data
#>       [,1]         [,2]        [,3]       [,4]        [,5]        [,6]
#>  [1,]    1  5.162165617  3.51781068 -0.4733455 -0.39681293  0.07974821
#>  [2,]    2  0.324331233  1.15199039  0.4501548 -0.34915681 -0.05706733
#>  [3,]    3  0.628674056  0.38842370  0.5037310 -0.30150070 -0.19388286
#>  [4,]    4  0.465073948 -0.32097173 -0.1552386 -0.25384458 -0.33069840
#>  [5,]    5  0.301473839 -2.39135311 -0.2304656 -0.20618847 -0.46751394
#>  [6,]    6  1.790640499  0.89829813 -0.3056925 -0.15853236 -0.60432948
#>  [7,]    7 -0.005856082 -0.11913019 -0.3809195 -0.11087624 -0.74114502
#>  [8,]    8  0.530457470 -1.81122400 -0.4561464 -0.06322013 -0.87796055
#>  [9,]    9 -1.112591695  0.35752554 -1.0144646 -0.01556402 -1.01477609
#> [10,]   10  0.006758480 -0.09005252 -0.4444690  0.03209210 -1.15159163
#>             [,7]       [,8]       [,9]    [,10]
#>  [1,] -1.2884072  0.4406636  1.1778013 1.994062
#>  [2,]  0.4744997 -1.0379310  1.9360774 2.497031
#>  [3,]  1.1574781  0.1665210 -0.4595392 3.000000
#>  [4,]  0.1428439 -0.6092184 -0.1425464 4.000000
#>  [5,]  0.1008368  0.5813857  0.4308495 5.000000
#>  [6,]  0.8609186 -0.2896958 -2.0217243 6.000000
#>  [7,]  2.5377739 -1.4210328  0.4704179 7.000000
#>  [8,]  0.4499884 -0.6940940  0.3414290 8.000000
#>  [9,]  1.9895189 -0.4123855 -1.0954659 9.000000
#> [10,]  1.2150913 -0.2301435  1.4910926 9.875742
#> 
#> $name
#> [1] "X"
#> 
#> $labels
#>  [1] "c.NA..NA..3.9..NA."                             
#>  [2] "c.NA..stats..rnorm.2...NA..stats..rnorm.6.."    
#>  [3] "stats..rnorm.10."                               
#>  [4] "c.stats..rnorm.4...rep.NA..3...stats..rnorm.3.."
#>  [5] "rep.NA..10."                                    
#>  [6] "c.stats..rnorm.1...rep.NA..9.."                 
#>  [7] "c.stats..rnorm.9...NA."                         
#>  [8] "stats..rnorm.10..1"                             
#>  [9] "stats..rnorm.10..2"                             
#> [10] "c.NA..NA..3.9..NA..1"                           
#> 
#> $intratime
#>  [1] 0.0000000 0.1111111 0.2222222 0.3333333 0.4444444 0.5555556 0.6666667
#>  [8] 0.7777778 0.8888889 1.0000000
#> 
#> attr(,"class")
#> [1] "dfts"