Functional KPSS Test

Compute the Kwiatkowski–Phillips–Schmidt–Shin (KPSS) statistic for functional data.

Usage

kpss_test(
  X,
  method = c("simulation", "resample"),
  resample_blocks = "separate",
  M = 1000,
  blocksize = 2 * ncol(X)^(1/5),
  TVE = 1,
  replace = TRUE
)

Arguments

X

A dfts object or data which can be automatically converted to that format. See dfts().

method

String for the method in computing thresholds: Monte Carlo simulation (simulation) or resampling (resample).

resample_blocks

String indicating the type of resample test to use. Using separate gives blocks which are separate while overlapping creates overlapping or sliding windows. When blocksize=1 then these will be identical.

M

Number of simulations to estimate theoretical distribution.

blocksize

Numeric for the block size when using a resample test.

TVE

Numeric for pca() to select the number of principle components.

replace

Boolean to indicate if blocks should be selected with replacement when using a resample test.

Value

List with the following elements:

pvalue: p-value from the test.

statistic: test statistic computed on the data.

simulations: Theoretical values for the null distribution.

References

Chen, Y., & Pun, C. S. (2019). A bootstrap-based KPSS test for functional time series. Journal of Multivariate Analysis, 174, 104535.

Kokoszka, P., & Young, G. (2016). KPSS test for functional time series. Statistics, 50(5), 957-973.

Examples

kpss_test(generate_brownian_motion(100,v=seq(0,1,length.out=20)))
#> $pvalue
#> [1] 0.934
#> 
#> $statistic
#> [1] 0.01464178
#> 
#> $simulations
#>    [1] 0.014748667 0.012062050 0.149000838 0.075622729 0.020529207 0.024508990
#>    [7] 0.016436551 0.061093280 0.026258436 0.041869018 0.021881223 0.018382314
#>   [13] 0.021909936 0.072500292 0.016891728 0.014178923 0.044097119 0.043453275
#>   [19] 0.022880455 0.015696774 0.016323938 0.058182421 0.040922858 0.018096475
#>   [25] 0.016725119 0.080160201 0.042213744 0.023886914 0.019926879 0.035599950
#>   [31] 0.030793800 0.026969008 0.024390064 0.019577569 0.033510834 0.016197621
#>   [37] 0.016309355 0.057237091 0.017803626 0.046074524 0.017011830 0.018466754
#>   [43] 0.061731224 0.050459909 0.042954003 0.024763488 0.028587187 0.039790701
#>   [49] 0.033857592 0.014265626 0.080968567 0.022270310 0.029964121 0.047118960
#>   [55] 0.034627735 0.057606835 0.024719854 0.019870412 0.025055180 0.017573888
#>   [61] 0.061979743 0.032210631 0.031443229 0.025341356 0.042705059 0.037799700
#>   [67] 0.020989976 0.057450130 0.012398871 0.044555933 0.031345093 0.023691119
#>   [73] 0.068400285 0.015727339 0.024791513 0.014360230 0.019034532 0.022721066
#>   [79] 0.016079037 0.018461544 0.026210997 0.049186728 0.023276706 0.019461844
#>   [85] 0.028932985 0.032056019 0.037317639 0.030567423 0.020612671 0.024097637
#>   [91] 0.069811243 0.029814779 0.045344089 0.012398691 0.031355445 0.012287685
#>   [97] 0.028010033 0.018036169 0.031199061 0.021430551 0.035567011 0.036920499
#>  [103] 0.060705432 0.016839849 0.085627694 0.045928139 0.033000980 0.061375212
#>  [109] 0.030609559 0.028523683 0.017884369 0.061985953 0.014241590 0.091303305
#>  [115] 0.012674689 0.031675535 0.020857391 0.037535350 0.075799049 0.014133508
#>  [121] 0.041602014 0.028364699 0.016150644 0.059529965 0.019952016 0.026137734
#>  [127] 0.036703730 0.059636958 0.032135785 0.032883132 0.035108515 0.049279918
#>  [133] 0.032378524 0.048877592 0.018951050 0.050371654 0.024776667 0.031838814
#>  [139] 0.067014044 0.018334646 0.040681856 0.060963541 0.026081863 0.032607707
#>  [145] 0.022637251 0.049363878 0.024580085 0.016989976 0.016041009 0.030814505
#>  [151] 0.027612580 0.030380992 0.043659901 0.019819000 0.020153801 0.012415829
#>  [157] 0.040624389 0.031027292 0.023859803 0.020105364 0.045569322 0.019315321
#>  [163] 0.023748773 0.036616967 0.026650062 0.025383186 0.050795276 0.040015142
#>  [169] 0.021992723 0.029837887 0.032504050 0.012752823 0.026229062 0.033593815
#>  [175] 0.031531378 0.042677900 0.018723067 0.037503140 0.025681182 0.060974102
#>  [181] 0.039587453 0.061278060 0.033720459 0.020402689 0.026427422 0.084075385
#>  [187] 0.034001321 0.014374891 0.049646591 0.026106694 0.029762182 0.030345978
#>  [193] 0.027392105 0.016422276 0.015578489 0.023358267 0.049177572 0.017517490
#>  [199] 0.018264928 0.029904575 0.013726020 0.037041101 0.033862722 0.015062529
#>  [205] 0.043220359 0.019717392 0.032234982 0.046126675 0.065007303 0.023490884
#>  [211] 0.026370989 0.032055746 0.026269770 0.011853459 0.053492771 0.029176299
#>  [217] 0.063545218 0.026850511 0.055756039 0.027060192 0.026289814 0.022120787
#>  [223] 0.086108299 0.013238876 0.024940637 0.046487435 0.007904297 0.018232056
#>  [229] 0.023687405 0.048614005 0.079230060 0.020122504 0.013320044 0.031027333
#>  [235] 0.022243107 0.059567077 0.025785897 0.025298698 0.051612809 0.040394757
#>  [241] 0.032330096 0.021538766 0.050166178 0.017926757 0.035823482 0.024959327
#>  [247] 0.016216635 0.033409689 0.010113539 0.030097392 0.027144049 0.045830936
#>  [253] 0.066611333 0.024430288 0.012281541 0.023135055 0.041264064 0.029310584
#>  [259] 0.042171012 0.053630092 0.043054617 0.036146499 0.229775500 0.123004536
#>  [265] 0.035504270 0.013817309 0.024643723 0.030638008 0.027507145 0.024532600
#>  [271] 0.032580470 0.046600729 0.015352817 0.064898276 0.023587173 0.048283745
#>  [277] 0.030071672 0.019179664 0.031037511 0.027426242 0.070899502 0.035700748
#>  [283] 0.031140933 0.028156241 0.090977637 0.061677003 0.048020131 0.012700132
#>  [289] 0.019280535 0.050927857 0.030410713 0.036207525 0.056713942 0.029246680
#>  [295] 0.054511306 0.022545716 0.018846883 0.021183718 0.028959842 0.040794664
#>  [301] 0.063791963 0.036347921 0.024694497 0.030479713 0.012745127 0.022544563
#>  [307] 0.038139242 0.013664132 0.056329619 0.016064733 0.015684868 0.018384777
#>  [313] 0.017665484 0.019983881 0.015645937 0.058705987 0.037190722 0.032626560
#>  [319] 0.033087922 0.017941241 0.089880101 0.043919541 0.021349435 0.024445635
#>  [325] 0.012895131 0.024118873 0.067222135 0.007199054 0.026282670 0.026833463
#>  [331] 0.026428069 0.050357485 0.054362549 0.024249485 0.035324158 0.030927640
#>  [337] 0.020150798 0.016438994 0.013648769 0.057992552 0.027280435 0.043746975
#>  [343] 0.047177068 0.052453556 0.044489154 0.017803375 0.040358896 0.041172127
#>  [349] 0.030401259 0.025713638 0.021742130 0.030208177 0.048743913 0.021048902
#>  [355] 0.029920951 0.044710359 0.014630190 0.017208612 0.058195215 0.040392845
#>  [361] 0.032245337 0.036987571 0.043666114 0.031608387 0.037105570 0.046867077
#>  [367] 0.035817587 0.029016903 0.032115718 0.022450934 0.024770348 0.022704901
#>  [373] 0.016660106 0.067649597 0.049761748 0.035740005 0.022083331 0.024620936
#>  [379] 0.029094310 0.027607655 0.014343506 0.023901206 0.026406597 0.014451195
#>  [385] 0.036610327 0.031750528 0.019654744 0.061874461 0.016631478 0.009638722
#>  [391] 0.017561676 0.056016403 0.018526909 0.029400157 0.021588614 0.023877867
#>  [397] 0.027383322 0.018626095 0.020568415 0.029360127 0.030061665 0.021928542
#>  [403] 0.023372529 0.021484436 0.026476001 0.020750777 0.214399325 0.018906568
#>  [409] 0.023864634 0.030897922 0.034685577 0.022988553 0.022944255 0.023616161
#>  [415] 0.030438222 0.055801679 0.011996748 0.063563460 0.018416718 0.029553388
#>  [421] 0.023235661 0.027880689 0.058219027 0.035843874 0.013208436 0.013472455
#>  [427] 0.015581389 0.029526734 0.016933017 0.035611112 0.034657632 0.073631112
#>  [433] 0.032265506 0.050040871 0.067503158 0.012260849 0.048356056 0.016156083
#>  [439] 0.069150385 0.029980418 0.012905983 0.022810849 0.021817399 0.016871546
#>  [445] 0.028292906 0.046205631 0.025988149 0.041294435 0.022338325 0.018639828
#>  [451] 0.028653625 0.023826041 0.036683699 0.017448348 0.021419637 0.028315385
#>  [457] 0.035570869 0.027141662 0.036478736 0.063856581 0.017075084 0.019803288
#>  [463] 0.012662676 0.013810315 0.026850430 0.059057154 0.021254327 0.064492229
#>  [469] 0.020508336 0.022678943 0.033843728 0.017038458 0.031897252 0.012761311
#>  [475] 0.020849816 0.037591760 0.031969312 0.018131062 0.040576110 0.019758394
#>  [481] 0.035468215 0.028383037 0.021995787 0.025370681 0.014919470 0.026997996
#>  [487] 0.062602734 0.017268697 0.019018351 0.022397197 0.034773455 0.041794000
#>  [493] 0.012476328 0.021160222 0.050502642 0.023915116 0.036563777 0.021062538
#>  [499] 0.016383334 0.021195204 0.062075981 0.025696431 0.051515164 0.031011659
#>  [505] 0.022903631 0.026913904 0.030463795 0.015721851 0.032157950 0.026670948
#>  [511] 0.019279362 0.032148506 0.046697040 0.039060289 0.021605212 0.016531600
#>  [517] 0.028988331 0.020073111 0.026965568 0.021544047 0.021870813 0.037646809
#>  [523] 0.035840732 0.033857821 0.059285437 0.070678749 0.046008632 0.026443751
#>  [529] 0.039469450 0.022785800 0.031315598 0.017822179 0.020634981 0.018437431
#>  [535] 0.025553773 0.018070071 0.029687577 0.032898781 0.018601999 0.030059186
#>  [541] 0.051553717 0.012871356 0.032205526 0.030921801 0.022720911 0.041957471
#>  [547] 0.033061640 0.030250986 0.037431357 0.048599668 0.066625919 0.083713896
#>  [553] 0.015349611 0.015486395 0.024617493 0.028019521 0.027953325 0.028069444
#>  [559] 0.020289103 0.024162071 0.043630277 0.014908272 0.018482213 0.008896665
#>  [565] 0.036300065 0.028652864 0.042570087 0.023038360 0.031077258 0.033322367
#>  [571] 0.012823463 0.060593483 0.005394051 0.044692143 0.016884806 0.028984990
#>  [577] 0.029437636 0.029969917 0.019252092 0.035003710 0.034195376 0.027085943
#>  [583] 0.043680646 0.019682380 0.014836684 0.045123153 0.054013102 0.034488190
#>  [589] 0.072265350 0.024604665 0.088158499 0.016561609 0.087119874 0.035900738
#>  [595] 0.027815987 0.058272374 0.086406221 0.047602568 0.041432652 0.025513093
#>  [601] 0.047727006 0.026396843 0.020075323 0.016173158 0.019680212 0.044066103
#>  [607] 0.026052570 0.018751035 0.085455186 0.032528996 0.039290638 0.063655181
#>  [613] 0.021107433 0.030124519 0.019776744 0.020869172 0.090323599 0.022289551
#>  [619] 0.020601513 0.036569912 0.059477253 0.044992437 0.032936784 0.033299634
#>  [625] 0.051052728 0.019817194 0.022351715 0.026908766 0.061509869 0.024675192
#>  [631] 0.021271187 0.047567152 0.014393714 0.032842573 0.022134558 0.029634174
#>  [637] 0.013197656 0.018108344 0.023292752 0.022422312 0.030046281 0.054809838
#>  [643] 0.023066042 0.028133962 0.022464223 0.022832913 0.020334632 0.031001007
#>  [649] 0.047122593 0.033119235 0.021578867 0.019295177 0.023848634 0.040452360
#>  [655] 0.024426270 0.037360295 0.031310004 0.040976456 0.021497880 0.034263535
#>  [661] 0.050367253 0.019486024 0.029170302 0.017222626 0.029833698 0.035986550
#>  [667] 0.042305156 0.026532317 0.031744080 0.023910158 0.025204348 0.023862074
#>  [673] 0.035983995 0.025937670 0.034979182 0.033744930 0.022454551 0.020973276
#>  [679] 0.038820967 0.051463403 0.076615956 0.023952035 0.022978605 0.022498002
#>  [685] 0.022366278 0.014453391 0.036970556 0.023029640 0.043897160 0.027279939
#>  [691] 0.036742344 0.042066790 0.034231408 0.020942552 0.048717770 0.012764972
#>  [697] 0.033003400 0.011999707 0.019090787 0.018573681 0.051623866 0.024253988
#>  [703] 0.038686092 0.065411066 0.041968670 0.025865835 0.025630881 0.020779370
#>  [709] 0.023920389 0.035043754 0.014223490 0.054290657 0.018944514 0.019640332
#>  [715] 0.041281961 0.021975913 0.028090848 0.022980819 0.012177980 0.036770999
#>  [721] 0.010243016 0.011780499 0.025605028 0.032232612 0.025972480 0.037709018
#>  [727] 0.048014174 0.017518129 0.015797818 0.016656083 0.014082859 0.021139875
#>  [733] 0.031266741 0.030550339 0.041660414 0.037311534 0.043595809 0.024969088
#>  [739] 0.020373519 0.035455763 0.016988029 0.020303876 0.046582569 0.035984673
#>  [745] 0.017425276 0.037123543 0.021319616 0.015634417 0.032044332 0.035788621
#>  [751] 0.043459440 0.033580039 0.029221210 0.033884817 0.054414113 0.031083838
#>  [757] 0.026870874 0.028402253 0.014835433 0.015568500 0.022643010 0.107458103
#>  [763] 0.019339443 0.020959002 0.048403436 0.011534593 0.092319834 0.023501053
#>  [769] 0.068778793 0.045281805 0.035312362 0.059375092 0.019211176 0.012962709
#>  [775] 0.015275945 0.061581985 0.028787929 0.048286433 0.051197967 0.072875944
#>  [781] 0.030030631 0.026094134 0.023461003 0.036504674 0.039308494 0.035297815
#>  [787] 0.014151597 0.061656839 0.037517440 0.031001589 0.026886618 0.042013409
#>  [793] 0.029552912 0.036449606 0.018631286 0.023049100 0.045251976 0.026295063
#>  [799] 0.030618299 0.051606676 0.059382520 0.044694517 0.032960021 0.028185394
#>  [805] 0.044093785 0.028637801 0.084301141 0.017729949 0.028057287 0.027662535
#>  [811] 0.035318963 0.044380569 0.020888818 0.026347556 0.046408599 0.029049238
#>  [817] 0.049921198 0.034235072 0.069295951 0.022977093 0.022909565 0.071517342
#>  [823] 0.031208667 0.043279927 0.268570449 0.026417246 0.037486442 0.016600460
#>  [829] 0.024386709 0.042912057 0.034401717 0.049540910 0.019320773 0.019208006
#>  [835] 0.014969529 0.018289969 0.039528175 0.036780368 0.044335541 0.021674208
#>  [841] 0.037114527 0.017552270 0.036066732 0.035642752 0.021853065 0.039408098
#>  [847] 0.052829766 0.024195546 0.032546992 0.022622183 0.015307889 0.035729184
#>  [853] 0.028309039 0.027952681 0.064155340 0.041927650 0.025893772 0.016601229
#>  [859] 0.043165631 0.028004677 0.020732563 0.047034087 0.026004087 0.035174463
#>  [865] 0.016232698 0.016624171 0.027544607 0.026075610 0.022286540 0.031501572
#>  [871] 0.029575838 0.057738464 0.032673081 0.015329958 0.140439169 0.014222647
#>  [877] 0.018616343 0.062165595 0.027849188 0.030453033 0.022145173 0.022619226
#>  [883] 0.030223250 0.016345573 0.017756012 0.020267740 0.032217960 0.022162741
#>  [889] 0.042636386 0.040080852 0.017637702 0.103707734 0.013025773 0.021148686
#>  [895] 0.082322098 0.109243447 0.049883363 0.017183235 0.040112006 0.030417942
#>  [901] 0.020141412 0.015871812 0.024239797 0.040044954 0.070910780 0.041561747
#>  [907] 0.025768224 0.025232587 0.043668991 0.041959379 0.024813558 0.012262055
#>  [913] 0.055393199 0.039291743 0.013989662 0.061574246 0.020561398 0.018211421
#>  [919] 0.016824259 0.019127397 0.025901318 0.020868807 0.023288283 0.034445376
#>  [925] 0.058329348 0.012156583 0.033023819 0.015390041 0.023020287 0.027225706
#>  [931] 0.022153113 0.013043046 0.014100685 0.043465767 0.044957278 0.040378599
#>  [937] 0.016687058 0.095542560 0.033616465 0.016215855 0.023069206 0.028835284
#>  [943] 0.043871552 0.021218239 0.023054851 0.026598417 0.016411125 0.029453562
#>  [949] 0.038167115 0.015331926 0.022488546 0.014459027 0.016769471 0.025543410
#>  [955] 0.011447363 0.032474845 0.035179436 0.101067430 0.038882067 0.038037288
#>  [961] 0.031567025 0.022086735 0.023765289 0.019766887 0.038307008 0.018525326
#>  [967] 0.041080968 0.021001351 0.026220114 0.026671370 0.029218792 0.021405171
#>  [973] 0.020683975 0.024305718 0.022730195 0.030161246 0.033935249 0.021908310
#>  [979] 0.035108516 0.022326086 0.055451290 0.031680427 0.048936837 0.046677802
#>  [985] 0.024532030 0.019814301 0.037350634 0.079827678 0.054955468 0.035337074
#>  [991] 0.028159727 0.037897030 0.025277510 0.037532594 0.035835820 0.016947984
#>  [997] 0.017801715 0.027826624 0.014841630 0.026715812
#> 
kpss_test(generate_brownian_motion(100,v=seq(0,1,length.out=20)),
             method="resample")
#> $pvalue
#> [1] 0.819
#> 
#> $statistic
#> [1] 0.02431359
#> 
#> $simulations
#>    [1] 0.025204586 0.062967832 0.032543833 0.084996079 0.032855533 0.115253644
#>    [7] 0.037296677 0.063874955 0.036034250 0.088763478 0.029066744 0.041621389
#>   [13] 0.032105484 0.041385960 0.018009381 0.028771275 0.032096677 0.029081565
#>   [19] 0.053403989 0.053719965 0.047179072 0.026629823 0.033781667 0.023471157
#>   [25] 0.031036655 0.014600580 0.074245074 0.051132892 0.029331863 0.049540508
#>   [31] 0.062642981 0.038181302 0.084935607 0.030971196 0.041940548 0.054014362
#>   [37] 0.047505713 0.167332178 0.048052311 0.039876942 0.042410502 0.036042789
#>   [43] 0.045512570 0.025614174 0.067602887 0.049381107 0.133937672 0.077149848
#>   [49] 0.027400648 0.037639021 0.033000260 0.025635923 0.074391405 0.039514723
#>   [55] 0.015781072 0.043244101 0.032280977 0.014098072 0.027957238 0.062448081
#>   [61] 0.033205172 0.087562406 0.026290245 0.014520712 0.060082723 0.030701493
#>   [67] 0.050723994 0.031420145 0.053199422 0.059260491 0.121760263 0.038130314
#>   [73] 0.041547340 0.024173135 0.038215737 0.092394486 0.083848847 0.056427734
#>   [79] 0.054125215 0.050436325 0.069358024 0.100664457 0.023627301 0.043673093
#>   [85] 0.072981461 0.026350814 0.036291958 0.049288155 0.023816333 0.024818911
#>   [91] 0.022344288 0.022626755 0.061024036 0.022781109 0.138686900 0.056319790
#>   [97] 0.020216011 0.060956276 0.030314033 0.079466683 0.075541674 0.019122182
#>  [103] 0.023931285 0.018080735 0.028050558 0.033120805 0.034113791 0.025676831
#>  [109] 0.038076669 0.023109716 0.115214401 0.106467214 0.026698525 0.025509932
#>  [115] 0.030744413 0.018798811 0.029442196 0.038059727 0.099522194 0.027567790
#>  [121] 0.053184928 0.033597148 0.035461221 0.043027545 0.039295446 0.030792561
#>  [127] 0.045571612 0.032268336 0.041222223 0.152230627 0.024912622 0.037042343
#>  [133] 0.091485244 0.026875312 0.114432625 0.024775022 0.029541083 0.055496774
#>  [139] 0.037055941 0.027617539 0.031389109 0.032477351 0.041958980 0.020126394
#>  [145] 0.055380011 0.031141418 0.067006398 0.020365251 0.027285715 0.071581104
#>  [151] 0.022020795 0.053175658 0.071412989 0.099841943 0.046969403 0.020262620
#>  [157] 0.030303782 0.026913798 0.026628473 0.055776935 0.047270746 0.025891237
#>  [163] 0.034692512 0.033985542 0.018576852 0.064294867 0.093868268 0.058439283
#>  [169] 0.183373828 0.047944937 0.043650677 0.030182985 0.036285151 0.025036572
#>  [175] 0.089147722 0.048908750 0.026037504 0.029251784 0.031504565 0.019592894
#>  [181] 0.062385310 0.023257053 0.010003683 0.018253562 0.029967985 0.020072631
#>  [187] 0.040241850 0.032679553 0.022204138 0.028577282 0.059561858 0.048925843
#>  [193] 0.037403139 0.025826053 0.022976746 0.074684267 0.054354538 0.043107413
#>  [199] 0.083721749 0.047939975 0.075374334 0.021148480 0.024843795 0.040331281
#>  [205] 0.050786698 0.019513317 0.027024697 0.030583064 0.041744710 0.040482546
#>  [211] 0.034950765 0.059241113 0.029323726 0.023945056 0.017441722 0.042993420
#>  [217] 0.068452685 0.040598187 0.015808532 0.070942491 0.037956926 0.023734112
#>  [223] 0.011946775 0.042724654 0.028463802 0.075923056 0.030644955 0.036993937
#>  [229] 0.047170652 0.024431467 0.025258991 0.057974708 0.028290076 0.030558412
#>  [235] 0.039368194 0.031899174 0.033105748 0.028359392 0.033709987 0.056178376
#>  [241] 0.065249032 0.030165724 0.079067786 0.055107002 0.026809487 0.074096877
#>  [247] 0.014564423 0.019443365 0.020991196 0.086127054 0.054853874 0.032478140
#>  [253] 0.058496139 0.025444264 0.025415828 0.037485045 0.058252586 0.043062366
#>  [259] 0.034734232 0.032384057 0.043049328 0.069297127 0.052798992 0.037229958
#>  [265] 0.025076979 0.039089518 0.047649152 0.047740484 0.015880013 0.020553283
#>  [271] 0.024483097 0.046632582 0.087065792 0.018436475 0.041081810 0.058488003
#>  [277] 0.061971801 0.027587087 0.054660429 0.037244435 0.067206645 0.069774405
#>  [283] 0.039565108 0.031909318 0.068927352 0.024145775 0.025947863 0.058385843
#>  [289] 0.053741731 0.016654311 0.026432900 0.043193633 0.045975251 0.034810147
#>  [295] 0.034047026 0.045417870 0.047128248 0.057831965 0.102718196 0.017520898
#>  [301] 0.047132495 0.034018535 0.044809432 0.039472347 0.059851799 0.049282907
#>  [307] 0.029428800 0.058918782 0.060277504 0.087763009 0.054453488 0.037149761
#>  [313] 0.041286136 0.032502938 0.038995140 0.035615768 0.027507732 0.021262894
#>  [319] 0.048685041 0.029230975 0.033541296 0.020799346 0.027075068 0.013097800
#>  [325] 0.023241173 0.018275920 0.050897381 0.035076823 0.029915792 0.057887874
#>  [331] 0.037015195 0.031237666 0.037628168 0.018332011 0.038848036 0.038795802
#>  [337] 0.032126223 0.073682889 0.060116476 0.083618702 0.066330211 0.081225010
#>  [343] 0.042840167 0.111544106 0.037795100 0.019960124 0.038579787 0.088849110
#>  [349] 0.102329191 0.033182079 0.064056001 0.041823892 0.107234215 0.022650283
#>  [355] 0.027026253 0.034086279 0.018255523 0.056964753 0.026102604 0.023840779
#>  [361] 0.072649754 0.026139343 0.039603753 0.033807587 0.091094558 0.035109252
#>  [367] 0.032521344 0.020244155 0.032072698 0.039136165 0.069368166 0.014110423
#>  [373] 0.042365151 0.029107659 0.019131130 0.025235080 0.041805931 0.037828963
#>  [379] 0.080407409 0.027749999 0.060031129 0.040957914 0.034444446 0.074001308
#>  [385] 0.023593928 0.063573726 0.028762426 0.077447734 0.067279856 0.020174986
#>  [391] 0.025709375 0.018352442 0.054218430 0.023942754 0.031817677 0.038285168
#>  [397] 0.027841957 0.017204301 0.029321380 0.091459249 0.039664912 0.020325955
#>  [403] 0.065461351 0.038412939 0.116354416 0.036227701 0.024860972 0.032752350
#>  [409] 0.038311174 0.042182964 0.052718243 0.054522845 0.032835959 0.100267253
#>  [415] 0.017777137 0.094843064 0.028370540 0.065867034 0.025651742 0.032050220
#>  [421] 0.046895359 0.031431388 0.034802328 0.062117886 0.038003328 0.028124965
#>  [427] 0.026027904 0.046623213 0.024230064 0.036221898 0.097320314 0.017813286
#>  [433] 0.047121249 0.027698996 0.034102444 0.044388353 0.041150497 0.054565971
#>  [439] 0.019415620 0.017055729 0.103324717 0.026902675 0.059688139 0.033060471
#>  [445] 0.093157935 0.025380842 0.022421419 0.027334532 0.041878054 0.048021057
#>  [451] 0.042041863 0.036885350 0.038369357 0.028416446 0.034312733 0.043350930
#>  [457] 0.050231478 0.031731915 0.022292159 0.028885611 0.029172370 0.109969829
#>  [463] 0.057287429 0.049996726 0.029173987 0.051764839 0.018798040 0.068203071
#>  [469] 0.026971339 0.034013189 0.046485461 0.035392098 0.042075218 0.032812830
#>  [475] 0.092243643 0.066378709 0.016101296 0.050370631 0.022458288 0.057207229
#>  [481] 0.085349994 0.024527375 0.028067450 0.026442482 0.088358548 0.036933133
#>  [487] 0.101163190 0.050538828 0.066228556 0.019625378 0.122848482 0.061411883
#>  [493] 0.021085854 0.039697161 0.023028517 0.021369688 0.016939343 0.046351304
#>  [499] 0.026867719 0.045702552 0.021430775 0.028556999 0.035303401 0.026472548
#>  [505] 0.095569771 0.028719274 0.040870081 0.045590279 0.038697953 0.080120235
#>  [511] 0.077721395 0.055973252 0.044897180 0.028426183 0.053548362 0.046385498
#>  [517] 0.018590298 0.099712357 0.030355563 0.014092366 0.078908850 0.039511533
#>  [523] 0.028476034 0.037236002 0.009178179 0.031190235 0.028742095 0.024112683
#>  [529] 0.021868319 0.025754662 0.022596151 0.045876789 0.025918705 0.038129216
#>  [535] 0.058728017 0.017295176 0.088127872 0.020005161 0.015871591 0.023512615
#>  [541] 0.026556097 0.039908124 0.047052382 0.041624685 0.018351256 0.028224029
#>  [547] 0.102157582 0.074079102 0.080120002 0.055573597 0.063098712 0.022965914
#>  [553] 0.042461634 0.070357410 0.018764935 0.025964091 0.015413432 0.032138821
#>  [559] 0.037437882 0.029061715 0.037289396 0.042180058 0.065258759 0.040756484
#>  [565] 0.029057734 0.043070350 0.069439522 0.025294205 0.098811551 0.063904059
#>  [571] 0.032547862 0.025190365 0.018138502 0.038678618 0.061179720 0.037430560
#>  [577] 0.045099006 0.101674025 0.046228239 0.048688614 0.052904415 0.051054703
#>  [583] 0.062021512 0.022534394 0.129828028 0.013821886 0.092719896 0.054154089
#>  [589] 0.049375217 0.026009936 0.011074582 0.025173819 0.034595223 0.027158613
#>  [595] 0.029700478 0.036692090 0.021657651 0.045597652 0.032509899 0.035023129
#>  [601] 0.016126742 0.053222446 0.040702903 0.022783653 0.028517147 0.026581990
#>  [607] 0.018957830 0.011718988 0.038244486 0.060585806 0.022688178 0.057217881
#>  [613] 0.039209759 0.055315194 0.026693219 0.039413009 0.031684176 0.020718143
#>  [619] 0.047501801 0.053276120 0.033149790 0.015738813 0.069742358 0.082732444
#>  [625] 0.045859050 0.020892551 0.015496478 0.077577708 0.022782916 0.025706388
#>  [631] 0.088424117 0.032876420 0.057121650 0.024451830 0.080907127 0.027028510
#>  [637] 0.031150084 0.025573830 0.022128563 0.025137983 0.038152341 0.030574781
#>  [643] 0.042855141 0.028682921 0.050462994 0.020379219 0.030842719 0.087456018
#>  [649] 0.020466519 0.051492955 0.061526356 0.057559526 0.014916324 0.034115795
#>  [655] 0.033531792 0.044225268 0.013144802 0.038995775 0.029256988 0.029894598
#>  [661] 0.035814739 0.018286103 0.019550584 0.054658252 0.022926013 0.029431155
#>  [667] 0.049047557 0.038972278 0.106696496 0.051867656 0.057078186 0.055055973
#>  [673] 0.033289636 0.028076840 0.024254368 0.047586388 0.036946611 0.030714517
#>  [679] 0.028588941 0.027336541 0.024752857 0.049690936 0.023690144 0.032462998
#>  [685] 0.029118248 0.024342096 0.023537077 0.027217605 0.018433623 0.059099968
#>  [691] 0.035091377 0.025120700 0.056199920 0.026479854 0.044196057 0.028078507
#>  [697] 0.025917273 0.014171409 0.056676988 0.060930103 0.039740363 0.029263318
#>  [703] 0.038523470 0.017471662 0.025752670 0.036859551 0.031203339 0.034224097
#>  [709] 0.042259040 0.019572652 0.026100271 0.035596408 0.033883032 0.019269083
#>  [715] 0.034714819 0.038963632 0.064047487 0.019663932 0.050907082 0.057100133
#>  [721] 0.031160734 0.033632225 0.061904780 0.018977827 0.027347106 0.024788166
#>  [727] 0.020942324 0.041030293 0.121572602 0.033879261 0.045284794 0.036901732
#>  [733] 0.102087306 0.030990999 0.035345916 0.023081550 0.009251662 0.054496024
#>  [739] 0.024732229 0.029708571 0.022174278 0.033663239 0.035480023 0.034996180
#>  [745] 0.063009015 0.074699453 0.098383098 0.030154557 0.050366435 0.033717307
#>  [751] 0.029756351 0.029232411 0.021039452 0.031668338 0.027043695 0.031287516
#>  [757] 0.017066415 0.023488961 0.019869862 0.027437923 0.017840352 0.016486707
#>  [763] 0.030535704 0.035633519 0.047627335 0.047814774 0.030458657 0.020987284
#>  [769] 0.025786927 0.063708221 0.025809229 0.016705713 0.034442766 0.027753490
#>  [775] 0.024147434 0.057803169 0.031156165 0.049781467 0.030851462 0.031774268
#>  [781] 0.034392040 0.033168441 0.026957165 0.048825471 0.020708858 0.058234970
#>  [787] 0.053859249 0.020261886 0.017937132 0.021151325 0.049467238 0.027830925
#>  [793] 0.025023132 0.022575484 0.117057434 0.050362518 0.041070064 0.061360178
#>  [799] 0.051090985 0.113871150 0.042019379 0.097137426 0.097069958 0.035153109
#>  [805] 0.029333057 0.030971588 0.065180630 0.103675753 0.022575471 0.032187785
#>  [811] 0.057718438 0.042930147 0.040218234 0.037981830 0.027267424 0.026533584
#>  [817] 0.018844853 0.013979539 0.027001604 0.040925731 0.048233671 0.022930613
#>  [823] 0.015046771 0.048579682 0.035879611 0.017392049 0.025435683 0.032040100
#>  [829] 0.046465457 0.037314970 0.044996491 0.012821828 0.063205000 0.036115988
#>  [835] 0.035829367 0.042034422 0.019197918 0.025692488 0.017584227 0.129722562
#>  [841] 0.052982963 0.025585397 0.047571088 0.062508938 0.082241110 0.030926538
#>  [847] 0.023048641 0.022375950 0.035972855 0.077808663 0.035569158 0.014172033
#>  [853] 0.035666070 0.061299398 0.049329664 0.047564039 0.034770444 0.048714914
#>  [859] 0.017783868 0.046587068 0.052248661 0.074801710 0.063552613 0.115802882
#>  [865] 0.035169035 0.026294399 0.089954556 0.047454389 0.036447019 0.033482339
#>  [871] 0.036992956 0.025394914 0.026255792 0.033470091 0.071919167 0.028927940
#>  [877] 0.013312613 0.041348459 0.036429074 0.032511915 0.063576383 0.056312551
#>  [883] 0.071334337 0.031956388 0.088486725 0.078926224 0.079559904 0.066061633
#>  [889] 0.035368477 0.057693254 0.026541587 0.042298393 0.055506388 0.042330622
#>  [895] 0.042116235 0.043043781 0.042284200 0.036024381 0.032266843 0.028104953
#>  [901] 0.045927611 0.078999526 0.071505744 0.026968437 0.021105467 0.095652839
#>  [907] 0.047900912 0.032845691 0.087990392 0.055089203 0.032003011 0.049839740
#>  [913] 0.031165779 0.058527175 0.037351567 0.083710681 0.019861713 0.027301918
#>  [919] 0.038742804 0.033609263 0.032900578 0.021536288 0.043862235 0.049395420
#>  [925] 0.046474101 0.021249865 0.051278707 0.036887708 0.043998019 0.026913876
#>  [931] 0.028278522 0.032954722 0.049558824 0.028280508 0.067286535 0.023894187
#>  [937] 0.026971474 0.041721160 0.038094426 0.040874658 0.029568478 0.025205581
#>  [943] 0.015693048 0.016404131 0.027291206 0.028308368 0.024434916 0.055396183
#>  [949] 0.041403959 0.020354813 0.017197387 0.011111207 0.022092002 0.040556550
#>  [955] 0.028990458 0.025809031 0.026966725 0.022750381 0.052880791 0.062939229
#>  [961] 0.009868520 0.041193386 0.034611795 0.019563701 0.020133382 0.032862363
#>  [967] 0.040013458 0.056978090 0.040271645 0.048956981 0.043555937 0.058616956
#>  [973] 0.043300022 0.037176277 0.043907084 0.045147855 0.030441486 0.020647118
#>  [979] 0.017415715 0.024563509 0.071429738 0.031195087 0.044506961 0.018981663
#>  [985] 0.015417145 0.018526218 0.030673485 0.026838480 0.067904193 0.037750050
#>  [991] 0.068058566 0.032221226 0.036118662 0.028817107 0.021389016 0.057000157
#>  [997] 0.058667655 0.030217941 0.053840326 0.058996392
#> 
kpss_test(generate_brownian_motion(100,v=seq(0,1,length.out=20)),
             method='resample', resample_blocks='overlapping')
#> $pvalue
#> [1] 0.32
#> 
#> $statistic
#> [1] 0.03689119
#> 
#> $simulations
#>    [1] 0.041486408 0.038210122 0.037017946 0.050796436 0.030589406 0.068038387
#>    [7] 0.030605880 0.029475749 0.035175852 0.072738079 0.033868342 0.024132149
#>   [13] 0.021364691 0.024594553 0.080376805 0.048732118 0.020922600 0.050917490
#>   [19] 0.020149384 0.054092007 0.022624239 0.016335947 0.027397084 0.026413512
#>   [25] 0.028277564 0.019936170 0.031329359 0.031099288 0.024794200 0.044892388
#>   [31] 0.031844030 0.034318819 0.049944471 0.033245991 0.029689528 0.041163218
#>   [37] 0.022473792 0.014491640 0.013048830 0.022183800 0.020007895 0.023936255
#>   [43] 0.018559494 0.035281561 0.047040258 0.025413910 0.057076513 0.017754586
#>   [49] 0.017297666 0.018311687 0.048545478 0.029064440 0.019727664 0.023904597
#>   [55] 0.016266816 0.015464308 0.021706527 0.024371532 0.019773563 0.035068616
#>   [61] 0.017355007 0.021804879 0.061868178 0.027099676 0.029083450 0.015743023
#>   [67] 0.059662500 0.027553549 0.032684669 0.022263335 0.028693187 0.055099137
#>   [73] 0.013914892 0.029367435 0.023090719 0.056140370 0.018338507 0.035648105
#>   [79] 0.102701331 0.030771937 0.077576142 0.018714298 0.025471635 0.061071151
#>   [85] 0.030260619 0.030546917 0.056446362 0.086716969 0.016107423 0.054948384
#>   [91] 0.028064934 0.031880361 0.046631456 0.029678387 0.040038778 0.053136176
#>   [97] 0.023858348 0.052745287 0.030022060 0.038014492 0.049644383 0.045582353
#>  [103] 0.052778680 0.072755649 0.062681458 0.044319039 0.030265795 0.032114963
#>  [109] 0.022578308 0.022551221 0.021775249 0.019607173 0.035517672 0.028049299
#>  [115] 0.035093981 0.018964709 0.059921298 0.053226096 0.030418056 0.031293328
#>  [121] 0.024453879 0.022140190 0.019325627 0.024976428 0.029676968 0.016661843
#>  [127] 0.070827630 0.032268750 0.013437029 0.062435868 0.061218409 0.031421157
#>  [133] 0.056561654 0.030394461 0.046550803 0.020211992 0.021143552 0.015638815
#>  [139] 0.045057308 0.011312316 0.045416115 0.066791100 0.030845123 0.033445790
#>  [145] 0.060672027 0.033970182 0.026303829 0.023935406 0.033727015 0.031718544
#>  [151] 0.036781577 0.014094200 0.034363949 0.013092134 0.032732037 0.025572610
#>  [157] 0.051032397 0.027060579 0.019760532 0.044988384 0.023397557 0.020559620
#>  [163] 0.042335900 0.027892264 0.027826745 0.021945687 0.041076005 0.015480930
#>  [169] 0.113932147 0.035045124 0.035857048 0.025894482 0.021381371 0.023944326
#>  [175] 0.038233564 0.037759135 0.018019702 0.043913488 0.024898525 0.059223573
#>  [181] 0.017325298 0.015171155 0.037126312 0.025244073 0.048460820 0.030310407
#>  [187] 0.031729452 0.047921848 0.051445323 0.042522911 0.051712955 0.071440463
#>  [193] 0.071009663 0.027125102 0.031839343 0.042832889 0.018876411 0.017408677
#>  [199] 0.027157160 0.047048148 0.010731788 0.021730218 0.029994395 0.048861929
#>  [205] 0.031880543 0.067893859 0.035307883 0.020260892 0.014479297 0.024195633
#>  [211] 0.014413003 0.015286363 0.030173499 0.039795511 0.027646195 0.025191378
#>  [217] 0.022575505 0.058110324 0.017198869 0.017437461 0.025112267 0.015382927
#>  [223] 0.016685588 0.034236496 0.041083052 0.045438160 0.020856096 0.025306981
#>  [229] 0.037264189 0.021262562 0.016034028 0.028600636 0.074609421 0.018017466
#>  [235] 0.061242406 0.043249589 0.025548229 0.013537532 0.012908070 0.027324746
#>  [241] 0.017226855 0.026009427 0.039439802 0.041922673 0.094114546 0.091052251
#>  [247] 0.035646503 0.044874308 0.052631953 0.027640391 0.039028932 0.016332190
#>  [253] 0.019036733 0.034465593 0.025029895 0.029153705 0.013774160 0.031740619
#>  [259] 0.028259409 0.034396512 0.039292181 0.028557373 0.066156351 0.009634126
#>  [265] 0.027758188 0.031161628 0.023606307 0.057161618 0.042733573 0.018637353
#>  [271] 0.051836575 0.013507565 0.036839465 0.045375864 0.032891225 0.023672549
#>  [277] 0.023450510 0.018484365 0.016657955 0.015894956 0.044462439 0.031039829
#>  [283] 0.019481577 0.017231395 0.026979628 0.024905964 0.024001734 0.021794305
#>  [289] 0.021918999 0.079004548 0.017135116 0.017835779 0.027765675 0.023713038
#>  [295] 0.017693825 0.029607768 0.046128369 0.036471396 0.057586997 0.025527803
#>  [301] 0.018026583 0.026431316 0.031521281 0.028830619 0.055976399 0.008672314
#>  [307] 0.036078234 0.025376455 0.080793912 0.027586062 0.019351709 0.031082016
#>  [313] 0.018332619 0.030333782 0.041944679 0.012532630 0.041405334 0.054507176
#>  [319] 0.036154035 0.025873193 0.028274959 0.031840659 0.023184250 0.023679180
#>  [325] 0.022021469 0.089022741 0.014666466 0.041661660 0.035527297 0.037268868
#>  [331] 0.036912110 0.029445034 0.014002774 0.024499840 0.025997656 0.034140806
#>  [337] 0.026180942 0.019825072 0.022970663 0.035375442 0.021115219 0.019391467
#>  [343] 0.025609661 0.018966711 0.037646501 0.015784618 0.038642568 0.027846232
#>  [349] 0.047112639 0.029651366 0.048814403 0.019566408 0.028364384 0.044347758
#>  [355] 0.027220576 0.018005237 0.037023973 0.033821135 0.021185517 0.052307617
#>  [361] 0.014180548 0.019977478 0.031320453 0.042695096 0.021068024 0.019067392
#>  [367] 0.058048433 0.022116530 0.032171191 0.016751303 0.022569422 0.129872448
#>  [373] 0.033250471 0.018277576 0.039612434 0.015466119 0.052899642 0.060976792
#>  [379] 0.029431797 0.051151157 0.053811209 0.015294998 0.023186021 0.030880582
#>  [385] 0.063122360 0.029692816 0.024343798 0.015035908 0.022765868 0.027524726
#>  [391] 0.040945592 0.027073529 0.036916349 0.045730773 0.013509130 0.031850426
#>  [397] 0.024511581 0.029550559 0.024398695 0.035621665 0.021660451 0.032463163
#>  [403] 0.034663794 0.047689807 0.074054714 0.101045258 0.103725016 0.029998594
#>  [409] 0.051957727 0.065740405 0.020290119 0.024070329 0.030679682 0.043379829
#>  [415] 0.058498842 0.024461442 0.030526381 0.018958640 0.036205562 0.022219643
#>  [421] 0.041158021 0.015341562 0.045717799 0.017160417 0.058217599 0.016949881
#>  [427] 0.042382440 0.015287304 0.034966873 0.055596759 0.021704023 0.015921216
#>  [433] 0.010270745 0.012452591 0.025726280 0.015308423 0.017933559 0.028833220
#>  [439] 0.019812127 0.038502035 0.024863285 0.039573554 0.051520026 0.117359802
#>  [445] 0.083214899 0.023736963 0.022964030 0.031267207 0.015019260 0.021524924
#>  [451] 0.030585189 0.014630489 0.049261309 0.037352506 0.021398658 0.021805627
#>  [457] 0.022817916 0.046135244 0.036621849 0.022067244 0.024552682 0.019283547
#>  [463] 0.014845785 0.036789483 0.033441679 0.026108880 0.012074494 0.035085677
#>  [469] 0.043894200 0.025028674 0.035062428 0.029888845 0.027395517 0.033673375
#>  [475] 0.037234067 0.029549402 0.062029622 0.024678149 0.014914068 0.020611794
#>  [481] 0.045820174 0.017580310 0.025253620 0.024674278 0.033424244 0.019247079
#>  [487] 0.021922216 0.091856611 0.045098887 0.025128759 0.032765428 0.024058729
#>  [493] 0.054254635 0.034547384 0.045683080 0.024756858 0.027512533 0.031657650
#>  [499] 0.017017975 0.038209298 0.031886910 0.052666601 0.012935203 0.040704468
#>  [505] 0.017600100 0.028838401 0.021439607 0.016594722 0.014265222 0.021085876
#>  [511] 0.029238431 0.064143373 0.028929460 0.043490444 0.014312424 0.027602927
#>  [517] 0.028823758 0.020771868 0.012978676 0.075618252 0.025352372 0.020074748
#>  [523] 0.047933541 0.017055828 0.020768690 0.035244368 0.012297365 0.009348207
#>  [529] 0.016699233 0.027612522 0.019256037 0.038499730 0.070321717 0.024789271
#>  [535] 0.047718729 0.067211315 0.036217852 0.032569242 0.025059457 0.029153035
#>  [541] 0.023178807 0.037118597 0.046044512 0.048968793 0.023712853 0.091107412
#>  [547] 0.022190518 0.033279579 0.022146200 0.014199870 0.015267251 0.034614944
#>  [553] 0.026988080 0.028731971 0.031035105 0.038354097 0.036399583 0.030251166
#>  [559] 0.057496203 0.022293453 0.055430150 0.024077141 0.031311299 0.020030843
#>  [565] 0.028622756 0.095049743 0.021667672 0.025723986 0.080306321 0.096912472
#>  [571] 0.016649964 0.150064178 0.019740532 0.033189866 0.061656113 0.014590050
#>  [577] 0.020094882 0.049402321 0.029519821 0.046944077 0.041517950 0.024450348
#>  [583] 0.022005728 0.112606238 0.087528167 0.024034249 0.048168627 0.020267769
#>  [589] 0.025353879 0.012068530 0.071332599 0.019105548 0.022930865 0.058638485
#>  [595] 0.049225653 0.024289291 0.028814396 0.016247300 0.032400620 0.023511233
#>  [601] 0.049567053 0.040539924 0.040465619 0.021043395 0.020747978 0.013230349
#>  [607] 0.031364878 0.014000461 0.034891730 0.016446547 0.030461578 0.052483063
#>  [613] 0.013966795 0.017536498 0.033022853 0.017271989 0.010913325 0.013150721
#>  [619] 0.026723621 0.039724237 0.017042083 0.038219529 0.027731251 0.049904510
#>  [625] 0.018057084 0.042667315 0.007937342 0.048068854 0.106301006 0.024887260
#>  [631] 0.047025627 0.046566086 0.049934560 0.029941542 0.017613008 0.051940547
#>  [637] 0.010964466 0.035252663 0.017459558 0.024550628 0.032631671 0.029958453
#>  [643] 0.019605632 0.041369818 0.025742574 0.029681956 0.050221739 0.014256458
#>  [649] 0.036337885 0.018619365 0.037485421 0.027020938 0.024418180 0.030970021
#>  [655] 0.034827455 0.024450466 0.052456996 0.035954504 0.056629977 0.055399430
#>  [661] 0.019141469 0.018863302 0.026118120 0.031462768 0.034148362 0.015121303
#>  [667] 0.025037002 0.033164319 0.035578161 0.029812129 0.018010606 0.076426501
#>  [673] 0.024458675 0.014584274 0.017273868 0.013799134 0.022712294 0.043102787
#>  [679] 0.061076515 0.035526858 0.017197266 0.044332586 0.032984947 0.025293431
#>  [685] 0.030213085 0.050704990 0.033039218 0.047378072 0.023879167 0.038434436
#>  [691] 0.072381117 0.024629100 0.038430915 0.013939489 0.022118826 0.024580898
#>  [697] 0.094399117 0.034059203 0.024023131 0.066520427 0.029214536 0.018645963
#>  [703] 0.028130205 0.063431997 0.020035836 0.021154528 0.032971346 0.034435290
#>  [709] 0.021695350 0.035904584 0.037178554 0.020433073 0.021722487 0.036953151
#>  [715] 0.043285776 0.015588218 0.041964021 0.030582315 0.026803237 0.017277586
#>  [721] 0.053997900 0.023486685 0.014901300 0.040516151 0.088755440 0.027839211
#>  [727] 0.018576780 0.018597474 0.020514673 0.014540134 0.037146031 0.036785891
#>  [733] 0.054840005 0.017022774 0.028608431 0.026832638 0.043373668 0.025157722
#>  [739] 0.038311111 0.042218983 0.044844566 0.065369457 0.033152147 0.026515881
#>  [745] 0.011774975 0.022570050 0.038095755 0.024988737 0.018480399 0.031498495
#>  [751] 0.039631300 0.039436349 0.032617599 0.042282484 0.035432153 0.024266746
#>  [757] 0.022997522 0.046071366 0.071709693 0.058203064 0.060930750 0.051098506
#>  [763] 0.026993717 0.030932711 0.031719906 0.011225538 0.041847618 0.029385587
#>  [769] 0.045712849 0.040957884 0.029066919 0.114061271 0.027914056 0.040860745
#>  [775] 0.038132294 0.026594614 0.058793309 0.026600657 0.053879942 0.023746992
#>  [781] 0.024936824 0.028894185 0.011523917 0.049492502 0.018441763 0.024428259
#>  [787] 0.053689062 0.032547901 0.042782036 0.050483996 0.013174203 0.022256747
#>  [793] 0.039007815 0.018424660 0.038204027 0.027800776 0.023307690 0.026338682
#>  [799] 0.043329473 0.017618037 0.030547049 0.028711488 0.064046554 0.031038637
#>  [805] 0.020826689 0.045348912 0.025570273 0.018245794 0.016955692 0.029685693
#>  [811] 0.018467326 0.059436421 0.034776170 0.063802600 0.051409831 0.083089629
#>  [817] 0.032434112 0.092158012 0.018099636 0.036009971 0.031569964 0.026621389
#>  [823] 0.022584535 0.025547066 0.017191462 0.010761444 0.047147955 0.050317564
#>  [829] 0.046683147 0.023936137 0.041287254 0.022030055 0.021991272 0.012703463
#>  [835] 0.051069390 0.028953869 0.035793828 0.018851910 0.040238443 0.025827513
#>  [841] 0.019121886 0.018818795 0.014441957 0.026313372 0.053064560 0.025004280
#>  [847] 0.050794626 0.027558785 0.028081010 0.024399764 0.034928555 0.070130028
#>  [853] 0.056775178 0.024062806 0.097569294 0.037645915 0.021108503 0.024863230
#>  [859] 0.044360186 0.023542825 0.035730623 0.035163613 0.016501558 0.067410643
#>  [865] 0.017010790 0.017555702 0.015605485 0.014198922 0.025802629 0.085132772
#>  [871] 0.026310526 0.029083646 0.012991738 0.026853617 0.033814442 0.032144999
#>  [877] 0.022347856 0.049273535 0.033382108 0.074938106 0.025565452 0.032051301
#>  [883] 0.026936598 0.027646280 0.040019349 0.061807129 0.057153476 0.052673469
#>  [889] 0.056033483 0.053738177 0.029768933 0.027377757 0.025473253 0.031140698
#>  [895] 0.018357666 0.020818132 0.014860680 0.038507643 0.145169161 0.055349018
#>  [901] 0.023517558 0.041597751 0.019619740 0.039283027 0.029230484 0.027605568
#>  [907] 0.017002602 0.022760173 0.030988808 0.013431998 0.054369384 0.045776321
#>  [913] 0.024006648 0.013653279 0.020721813 0.042247888 0.031263845 0.035653171
#>  [919] 0.033059002 0.041718272 0.068211692 0.046351921 0.031816635 0.040278706
#>  [925] 0.017461792 0.012715165 0.051164445 0.017785637 0.027273834 0.024862390
#>  [931] 0.028465020 0.015454511 0.018668831 0.026091673 0.080201079 0.017448174
#>  [937] 0.100589953 0.048918835 0.017935610 0.078980284 0.027956066 0.036581212
#>  [943] 0.017084222 0.015891311 0.092815288 0.039530673 0.033805623 0.017186510
#>  [949] 0.039831229 0.031007719 0.013811858 0.087412318 0.024459635 0.034753688
#>  [955] 0.053495972 0.101543174 0.043498726 0.046822666 0.013015221 0.030135696
#>  [961] 0.031144869 0.093110420 0.031767272 0.017938831 0.028930324 0.020738109
#>  [967] 0.034556678 0.015924842 0.044929849 0.008531756 0.049905857 0.020106305
#>  [973] 0.042729731 0.033246228 0.044205954 0.049047838 0.049786605 0.054024162
#>  [979] 0.023628746 0.028128476 0.057214926 0.020722295 0.034735979 0.077325477
#>  [985] 0.021358463 0.018345172 0.030337535 0.022179452 0.027516916 0.069289265
#>  [991] 0.017514854 0.026404452 0.021059436 0.023208010 0.021921928 0.020312243
#>  [997] 0.017227628 0.021822235 0.087377752 0.019015345
#>