Estimate Long-run Covariance Kernel

This (internal) function estimates the long-run covariance kernel. That is, \(C_{\epsilon}(t,t') = \sum_{l=-\inf}^{\inf} \text{Cov}(\epsilon_0(t), \epsilon_l(t'))\) with error sequence \((\epsilon_i : i \in \mathbb{Z})\).

Usage

long_run_covariance(X, h = 0, K = bartlett_kernel)

Arguments

X

A dfts object or data which can be automatically converted to that format. See dfts().

h

The window parameter parameter for the estimation of the long run covariance kernel. The default value is h=0, i.e., it assumes iid data

K

(Optional) Function indicating the Kernel to use if h>0

Value

Data.frame of numerics with dim of ncol(data) x ncol(data), that is symmetric.

Details

This is an internal function, see use in .change_mean.

Examples

result <- long_run_covariance(electricity,2)