Generic Function to Compute QQ plot for dfts

Creates normal QQ plots on the principal components of functional data

Usage

# S3 method for class 'dfts'
qqplot(
  x,
  TVE = 0.95,
  d.max = 3,
  alpha = 0.05,
  changes = NULL,
  legend = FALSE,
  ...
)

Arguments

x

A dfts object or data which can be automatically converted to that format. See dfts().

TVE

Numeric in [0,1] giving the total variance explained for selecting the number of principal components.

d.max

Max number of principal components

alpha

Significance level, alpha in [0,1]

changes

Vector of change points

legend

Boolean indicating if legend should be shown

...

Unused

Value

ggplot2 for QQ plot

Examples

result <- qqplot(electricity)