Compute Spacing Measuring Functions

This function is used to compute discretized functions, i.e. vectors, to explore functional spaces.

Usage

space_measuring_functions(X, M = 20, space = "BM")

Arguments

X

A dfts object or data which can be automatically converted to that format. See dfts().

M

Integer for the number of functions to generate.

space

String for the space of interest. Options are Brownian motion ('BM'), OU process ('OU'), principal components ('PC'), and vectors in iid standard, random normals ('RN').

Value

Data.frame with columns of discretized functions describing the space. Columns are independent functions.

See also

Examples

space_measuring_functions(M=10, space="BM", X=electricity)
#>              [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#>  [1,]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [2,] -0.12825110  0.13263903  0.42028712  0.03826323  0.28338046  0.08073990
#>  [3,]  0.02941002  0.13765600  0.62075191 -0.04155553  0.02409807 -0.19228067
#>  [4,]  0.08234971  0.15981240  0.63822585  0.34137784 -0.33507473  0.11914744
#>  [5,]  0.07783359  0.13906444  0.78221831  0.30627969 -0.41178691 -0.06089779
#>  [6,]  0.13603079 -0.04905541  1.18050113  0.20815829 -0.75220385 -0.10211108
#>  [7,]  0.02633827 -0.17506061  0.75166316  0.04099872 -0.79221835 -0.25871766
#>  [8,]  0.06959954 -0.34202414  0.90344101 -0.32165420 -0.83382692 -0.09246447
#>  [9,]  0.10885845 -0.26551227  0.53442618 -0.58054122 -0.92382591  0.16467303
#> [10,]  0.21727245 -0.23516578  0.62093054 -0.58125481 -1.06349192  0.03825699
#> [11,]  0.15601845 -0.28070179  0.52767550 -0.60850433 -1.15514884  0.08239998
#> [12,]  0.29245716 -0.44647692  0.31842311 -0.45027979 -1.06546195 -0.42944076
#> [13,]  0.12011984 -0.56930639  0.46370036 -0.74914730 -1.20330691 -0.43227507
#> [14,]  0.09247954 -0.62572849  0.36099671 -0.96972711 -1.51347294 -0.23492731
#> [15,] -0.03059691 -0.47118896  0.29390402 -0.90326736 -1.39445426 -0.09662440
#> [16,] -0.06397907 -0.61118114  0.23199639 -1.00998926 -1.36293401  0.02357914
#> [17,] -0.12714004 -0.40575677  0.54794917 -0.90303410 -1.19733205 -0.24485502
#> [18,] -0.11534123 -0.05266264  0.41638728 -0.73689153 -1.02929778 -0.40001326
#> [19,] -0.37206229 -0.60307684  0.12758008 -0.64696881 -1.09402420 -0.64976188
#> [20,] -0.30068180 -0.70569295  0.20479969 -0.58607033 -0.88346402 -0.26436268
#> [21,] -0.43677067 -0.34345170 -0.01299151 -0.89516394 -0.57750824 -0.34662497
#> [22,] -0.57682258 -0.22819096 -0.26287339 -0.52576895 -0.52674725 -0.45159659
#> [23,] -0.69792797  0.19837177 -0.28699501 -0.20804524 -0.33956864 -0.43596715
#> [24,] -0.87951145  0.35315583 -0.48731516  0.19230348 -0.50622107 -0.43737263
#>               [,7]        [,8]        [,9]     [,10]
#>  [1,]  0.000000000  0.00000000  0.00000000 0.0000000
#>  [2,]  0.003524309 -0.07465334  0.31620525 0.2562833
#>  [3,]  0.246593675 -0.02108275  0.18307795 0.5560981
#>  [4,]  0.154220285 -0.17118446  0.13698431 0.5368694
#>  [5,]  0.116167782 -0.10525641  0.12891540 0.5596203
#>  [6,] -0.073372261 -0.24087792  0.24044500 0.5309715
#>  [7,] -0.169394295 -0.48269360  0.28370659 1.0192892
#>  [8,] -0.430879399 -0.80541578  0.06029931 1.0518934
#>  [9,] -0.606310382 -0.61449160  0.61701111 0.9310463
#> [10,] -0.593287373 -0.72716711  0.42656212 0.7265570
#> [11,] -0.672747807 -0.47746796  0.38697591 0.8629767
#> [12,] -0.765293053 -0.28589076  0.04267130 0.7486685
#> [13,] -0.755993028 -0.32807234 -0.08710088 1.0628633
#> [14,] -1.110695290 -0.03049371  0.07328941 1.2207766
#> [15,] -0.898440837  0.42952007  0.18296070 1.2705485
#> [16,] -1.103276229  0.17022759  0.11294254 1.2831768
#> [17,] -1.191173845 -0.03477435  0.15980911 1.4802667
#> [18,] -1.102678303 -0.31731649  0.19482726 1.3631990
#> [19,] -0.831893509 -0.47201629  0.13878652 1.5879056
#> [20,] -0.857086725 -0.38928996  0.71480642 1.6028035
#> [21,] -0.996801621 -0.57350606  0.77703806 1.4760782
#> [22,] -0.927160359 -0.32665618  0.53729754 1.5991292
#> [23,] -1.099739616 -0.21603597  0.55553696 2.0094545
#> [24,] -1.234089393 -0.46054406  0.49679703 1.8063915
#space_measuring_functions(M=10, space="OU", X=electricity)
space_measuring_functions(M=10, space="PC", X=electricity)
#>             V1          V2          V3         V4           V5           V6
#> 1   -346.01880   273.22612   767.35308  -650.9797  -406.091398 -1049.351718
#> 2  -1705.62869  -556.12931   667.61862  -567.2221   156.445373   337.100605
#> 3   -254.00380   215.55687   235.95398  -275.4357   -86.887905   597.988970
#> 4   -673.00468   270.37688  -497.99872 -1544.1558   -69.995296   781.200496
#> 5   -331.40671   870.31313  -426.38467  1025.2608  -808.591401 -1107.306543
#> 6     27.87124 -1255.82586   179.93952  1046.2258  -859.449766   742.100337
#> 7    249.26863   201.94510   -65.76114  1092.6840   177.467633 -1423.564241
#> 8    -95.43320  -709.37560   652.04333  -202.7654  -538.838779   119.245069
#> 9   -239.82954   -33.88626   393.49309  -132.4410   671.562291  -462.289679
#> 10 -2162.14572   672.73244   917.63987   453.2346  1741.063094  -910.212967
#> 11   401.27220  -534.84164   -83.46303   769.5926   420.442669  1018.671771
#> 12  1453.05787  -415.28951  1934.32929   606.2772   249.530811  -990.424866
#> 13  2567.69989  1117.01450  -459.93712   145.2534  -269.936082   362.142661
#> 14   177.42711   761.93286   836.60762  -124.7161   339.034894  -406.625865
#> 15  -138.89224  -655.20425  1751.63808  -864.2137 -1097.366913  1165.601620
#> 16   544.64293   597.19984  1104.50818  -295.6095    -2.889343  -450.282547
#> 17  -759.45882  -481.98574  1089.07606  -526.3823  -388.830453  -451.679308
#> 18 -1294.80592 -1270.47234   425.38856 -1147.8758 -1412.545802   610.663171
#> 19 -1351.38743  -149.55253  -760.82995   509.6685  -891.135738   899.347275
#> 20  -788.56997  -202.60745   498.69950   201.8065   836.567731  -889.383392
#> 21 -1168.01815  -732.85823  -489.32765  -603.0866   -72.370226   479.944790
#> 22   317.46063  -458.18981  -990.99422 -1036.2224  -261.309752    -9.985161
#> 23  -263.28547   292.55050    96.42579  -303.8265   843.735864  -354.250194
#> 24    30.72551    55.02851 -1178.21841  -348.7115   844.624046 -1298.107858
#>             V7          V8          V9         V10
#> 1   -763.46246   890.04671  1152.96377   998.67571
#> 2   -631.95544  -145.85551   150.49720   447.23312
#> 3   -693.79086   128.59396   485.32805  -443.06021
#> 4  -1625.36818  1072.47796  -577.93686 -1219.32895
#> 5  -1005.22235   684.69419  -916.44885 -1064.99208
#> 6   -568.38690  1221.94145   865.10692 -1040.47512
#> 7    412.10208   369.35869   940.94827  -377.11673
#> 8   -106.59867  2157.23542   594.66263   745.63736
#> 9   -526.38585  -398.31450  1246.29688    74.37631
#> 10   927.04824  -717.47096 -1686.22973  -633.82082
#> 11   437.90531 -1937.53307   248.37401   625.88673
#> 12  1326.37414   790.77039 -1721.34247  2108.10692
#> 13  -354.85394   756.96730   859.21855 -1698.01137
#> 14  -728.37979  1314.12894 -1174.93784  -181.58039
#> 15  -503.60390   709.78621  -808.52331  1346.73791
#> 16  1275.97203   119.24287   715.95893  1702.73139
#> 17  -611.75224   -45.09980   606.23691  -468.78849
#> 18   549.17770  -876.81925   195.61712   442.47123
#> 19  1051.28507 -1204.15284  -933.70239  -235.00103
#> 20   627.95367   -82.87919  -959.09825 -1661.79832
#> 21   267.92346   265.69191    46.68587  -613.84706
#> 22   -94.47965  1562.00865   241.02344  -401.75535
#> 23   -42.06652   122.85616  -132.86843  1607.82962
#> 24  1852.68673  -195.02309  -544.34229   170.09302