Compute Spacing Measuring Vectors

This function is used to compute vectors from spaces to measure the function data objects.

Usage

space_measuring_vectors(X, M = 20, space = "BM")

Arguments

X

A dfts object or data which can be automatically converted to that format. See dfts().

M

Integer for the number of vectors to generate

space

String for the space of interest. Options are Brownian motion ('BM'), OU process ('OU'), principal components ('PC'), and a vectors in iid standard, random normals ('RN').

Value

Data.frame with columns of vectors describing the space. Columns are independent vectors.

Examples

space_measuring_vectors(M=10, space="BM", X=electricity)
#>              [,1]          [,2]        [,3]        [,4]       [,5]         [,6]
#>  [1,]  0.00000000  0.0000000000  0.00000000  0.00000000  0.0000000  0.000000000
#>  [2,]  0.25418822 -0.2607752739  0.04351663 -0.01640834 -0.1551238 -0.050595264
#>  [3,]  0.23540426 -0.0110772794  0.17833894  0.22080534 -0.5408895  0.056055657
#>  [4,]  0.03535060 -0.0172864562  0.02478558  0.11254395 -0.6119310  0.478997749
#>  [5,] -0.08001589 -0.0684676211 -0.12190273 -0.35664470 -0.7283260  0.219487399
#>  [6,]  0.02842450 -0.1765245685 -0.02883417 -0.61765379 -0.5079469  0.084097185
#>  [7,]  0.20456242 -0.2439475076  0.08484219 -0.44666594 -0.8327540  0.029834281
#>  [8,]  0.36381951  0.0210952513 -0.19116022 -0.46219396 -0.8784770  0.430176123
#>  [9,]  0.29099723 -0.0771575492 -0.57034527 -0.58226696 -1.0674100  0.008783773
#> [10,]  0.31770187  0.2326972458 -0.66324112 -0.18689170 -1.2571463 -0.140018536
#> [11,]  0.27392581 -0.1065601456 -0.75304334 -0.51407573 -1.6120179 -0.583031286
#> [12,] -0.17961347 -0.0071913971 -0.82251599 -0.40237889 -1.8411637 -0.503773997
#> [13,] -0.37736262 -0.1979170860 -0.70370604 -0.44686029 -1.7020061 -0.174238155
#> [14,] -0.49284082 -0.1576687526 -0.95792147 -0.31050572 -1.9912691 -0.294143651
#> [15,] -0.63979045 -0.2591126749 -0.67178952 -0.42230234 -1.9939729 -0.607643132
#> [16,] -0.82614327 -0.1657275556 -0.72385872 -0.33149195 -1.3568731 -0.562134803
#> [17,] -0.79191473  0.0007771205 -0.98310109 -0.18619247 -1.6302181 -0.440690886
#> [18,] -0.68468507 -0.1643991549 -0.98338829 -0.19822982 -1.6936379  0.017346500
#> [19,] -0.71488530  0.1631092793 -1.18089865  0.25324158 -1.6334259  0.173619911
#> [20,] -0.84917925  0.5091521802 -1.45913673 -0.08618792 -1.7942783 -0.037527623
#> [21,] -0.39817871  0.2674695779 -1.87948708 -0.21129753 -1.7120912 -0.040799286
#> [22,] -0.91868535  0.1656004929 -1.92564203 -0.21313686 -1.5184470 -0.105881623
#> [23,] -0.98472202  0.1527352874 -2.17053563 -0.13212816 -1.2216554 -0.214123021
#> [24,] -1.40364669  0.2161304850 -1.94202138 -0.25160920 -0.9791805  0.415379389
#>              [,7]       [,8]        [,9]       [,10]
#>  [1,]  0.00000000  0.0000000  0.00000000  0.00000000
#>  [2,]  0.06755798 -0.1912902 -0.22884206  0.13116899
#>  [3,]  0.22410726 -0.1681280 -0.39614707 -0.58399971
#>  [4,]  0.46984051 -0.4783295 -0.04118298 -0.33782671
#>  [5,]  0.75368107 -0.8594931 -0.14441085 -0.09073257
#>  [6,]  0.44872388 -0.7554750 -0.12432593 -0.27451441
#>  [7,]  0.82503394 -0.9475307  0.14344703 -0.45954874
#>  [8,]  0.56727101 -0.9735186  0.14155632 -0.16608105
#>  [9,]  0.23447721 -1.1876742  0.08113525 -0.09497139
#> [10,]  0.33937115 -1.0940179  0.32232138 -0.15279576
#> [11,]  0.04461335 -1.0058343  0.23314489 -0.24886847
#> [12,]  0.16459791 -1.1801516  0.11600041 -0.11753188
#> [13,] -0.14952075 -1.0966103 -0.30404212  0.17167666
#> [14,]  0.07334185 -0.7601619 -0.38337087  0.48610620
#> [15,] -0.01290766 -0.3963934 -0.31092064  0.29466995
#> [16,] -0.13950320 -0.5258997 -0.21181241  0.63886264
#> [17,] -0.24577436 -0.3258882 -0.16804639  0.55582078
#> [18,] -0.26535005 -0.4527611 -0.24560912  0.66997651
#> [19,] -0.24275074 -0.7372388 -0.46577388  0.50164241
#> [20,] -0.19160458 -0.3792248 -0.22974327  0.25223371
#> [21,] -0.17746065 -0.3888268 -0.46628870  0.41386726
#> [22,] -0.11951040 -0.3117808 -0.56169408 -0.01702550
#> [23,] -0.14607569 -0.1426850 -0.59694329  0.15232493
#> [24,] -0.14641430 -0.3589049 -0.32979896  0.42309625
#space_measuring_vectors(M=10, space="OU", X=electricity)
space_measuring_vectors(M=10, space="PC", X=electricity)
#>            V1          V2          V3         V4          V5          V6
#> 1   377.15538 -1115.40545   -56.73571  -463.1231   950.45808  -377.54252
#> 2   950.32314  -365.26981  -846.45690  1162.9780  -159.41135  -491.01405
#> 3  1651.40626   309.67461   702.70959   351.7791   194.79148    21.52847
#> 4  -503.43600 -1304.22837 -2059.61572  -666.2209   141.55141  -524.44144
#> 5  -732.88719 -1940.15527   134.50044  -274.9304   438.17447   571.77302
#> 6  -505.89307  -697.24838   409.31559  -331.3790  1480.05897  -132.94424
#> 7  -525.53544  -580.56912   178.21749  -630.5537    37.46077   479.63849
#> 8  1246.68621 -1668.21607  -465.93979  -645.1443  -657.29136  -271.02319
#> 9  1016.11476  -343.47560 -1268.87491  -255.0774    -3.60259  -375.79592
#> 10 -105.61684  -358.61776  -508.40121  1035.7128  -835.10514 -1474.51831
#> 11  651.91285  2337.10838   946.51362   269.3005    41.06047   605.49438
#> 12  -64.77032  -608.03461 -1291.06706  -376.3419   220.88827    16.22337
#> 13 -984.99602 -1423.48818   124.56824  1170.6135  -682.51509   -50.02302
#> 14 -379.45070  -613.84515   971.70634  -344.0333  1016.82574   296.98645
#> 15 -190.75051 -1821.39401  -741.38833 -1134.7966  -164.41019  1009.94419
#> 16 -822.18140   906.82802  -917.89890  -284.1828    97.39430   590.37967
#> 17 -299.55848  -931.76629  1752.34846 -2489.2894  -681.65089  1643.72576
#> 18  616.09886   379.09201  -498.39297 -1975.8143    29.29593   773.48110
#> 19 1470.11095   568.44335 -1025.26218 -1503.9506   365.64816  1009.90317
#> 20 -485.31821   -25.60619   934.42043  -321.9652   -54.68179 -1412.85554
#> 21 -300.89478   374.21657  -790.64184 -1339.4291 -1694.68006   240.12023
#> 22  354.31327   197.73840  -432.85105  -186.2489   -65.47206    80.46541
#> 23  165.82814  -627.40208  1073.70128 -1496.6727 -1125.30440 -1753.32300
#> 24 -535.94750  -193.26890    74.99851 -1430.5559  -543.01252  2151.34175
#>             V7          V8         V9          V10
#> 1   -570.42667  -756.95928   690.0510   392.789149
#> 2   -535.11061  -488.01438  -710.5438 -1132.425671
#> 3   -257.81943  1198.72880   812.6882 -1005.736396
#> 4   -852.50917  -687.21520  -449.0622   209.626746
#> 5   -227.03224   -27.98216  -169.6400  -893.031989
#> 6   -570.71216   648.33199   742.3744     9.230325
#> 7   1010.17007   444.47535 -2250.9099  -152.347249
#> 8    242.00952   -51.02217   943.8104   324.248369
#> 9   -462.23211  -814.18616   646.3833  1581.037471
#> 10  -486.27490  3302.78701  -609.3656 -2275.852871
#> 11   915.48838  1131.06991  -763.7796  1177.772446
#> 12  -258.12464  1587.69999  -368.9911  1537.142392
#> 13   486.35123  -782.60410 -1137.2591  1520.365727
#> 14 -1183.21447  1387.08811  1220.9104  2136.141964
#> 15  -690.32152    42.77748   430.7832 -1932.977712
#> 16  1414.56495  1168.42643  -889.0324  1688.956989
#> 17   317.48454  -344.54991   241.0039  -828.710429
#> 18 -1596.53810   150.75575  -413.2870  1625.465612
#> 19  -195.20503  -983.53456  -689.9569   544.544311
#> 20  1076.91893  -291.35818   213.2337    80.384343
#> 21   -68.46426  -531.73932  1120.9490  1153.335675
#> 22   933.27529 -1040.25783   289.0451  -295.191939
#> 23  -491.03134 -1162.65158  -849.4873   214.184199
#> 24   579.69900   651.63166  1296.7133  -168.100206