Draw 3D Geoms for ggplot2

This function adds 3D geoms such as points and paths to a ggplot2 plot.

Usage

stat_3D(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

Arguments

mapping

Default list of aesthetic mappings to use for plot. If not specified, must be supplied in each layer added to the plot.

data

Default dataset to use for plot. If not already a data.frame, will be converted to one by fortify(). If not specified, must be supplied in each layer added to the plot.

geom

The geometric object to use to display the data for this layer. When using a stat_*() function to construct a layer, the geom argument can be used to override the default coupling between stats and geoms. The geom argument accepts the following:

A Geom ggproto subclass, for example GeomPoint.

A string naming the geom. To give the geom as a string, strip the function name of the geom_ prefix. For example, to use geom_point(), give the geom as "point".

For more information and other ways to specify the geom, see the layer geom documentation.

position

Set position information. Find more details in ggplot2 function.

na.rm

Boolean if na data should be removed.

show.legend

logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.

inherit.aes

If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders().

...

Arguments passed on to layer. Often the aesthetics like color = "red" or size = 3. Two important ones are theta (azimuthal rotation) and phi (colatitude rotation) to define angle in degrees of viewing data.

References

Acker D (2024). gg3D: 3D perspective plots for ggplot2. R package version 0.0.0.9000.

Examples

dat <- electricity

data_lines <- cbind(data.frame('Time'=dat$fparam), dat$data) %>%
tidyr::pivot_longer(cols = 1+1:ncol(dat$data))

colors_plot <- RColorBrewer::brewer.pal(11, "Spectral")
colors_plot <- grDevices::colorRampPalette(colors_plot)(ncol(dat$data))
data_lines$color <- rep(colors_plot, nrow(dat$data) )
data_lines$name <- as.numeric(data_lines$name)

result <- ggplot2::ggplot(data_lines,
  ggplot2::aes(y=Time, x=name, z=value, color=color)) +
  ggplot2::theme_void() +
  stat_3D(theta=0, phi=15, geom='path') +
  ggplot2::scale_color_manual(
     breaks = data_lines$color,
     values = data_lines$color
  ) +
  ggplot2::guides(color='none')