Functional Stationarity Test

Stationarity test for functional time series with different methods on determining the critical values of the test statistic. The Monte Carlo method was constructed in Horvath et al. (2014), while the bootstrap-based methods have not been validated in the literature (although such an option is provided, please use them at your own risk).

Usage

stationarity_test(
  X,
  statistic = "Tn",
  critical = c("simulation", "permutation"),
  perm_method = "separate",
  M = 1000,
  blocksize = 3,
  TVE = 1,
  replace = TRUE
)

Arguments

X

A dfts object or data which can be automatically converted to that format. See dfts().

statistic

String for test statistic. Options are Tn and Mn. Default is Tn.

critical

String for method of determining the critical values. Options are simulation and permutation. Default is simulation.

perm_method

String for method of permutation. Options are separate and sliding. Default is separate.

M

Numeric for number of simulation to use in determining the null distribution. Default is 1000.

blocksize

Numeric for blocksize in permutation test. Default is 1.

TVE

Numeric for total variance explained when using PCA for eigenvalues. Default is 1.

replace

Boolean if replacement should be used for permutation test. Default is TRUE.

Value

List with (1) pvalue, (2) test statistic, and (3) test statistics from simulating the null distribuion.

References

Horvath, L., Kokoszka, P., & Rice, G. (2014). Testing stationarity of functional time series. Journal of Econometrics, 179(1), 66-82.

Examples

res <- stationarity_test(
  generate_brownian_motion(100,v=seq(0,1,length.out=20)),
  critical='permutation', statistic='Mn')
res2 <- stationarity_test(electricity)