Generate Data via Ornstein-Uhlenbeck Process

Generate autoregressive process with errors according the Ornstein-Uhlenbeck process.

Usage

generate_ornstein_uhlenbeck(N, v, rho = 0)

Arguments

N

Numeric for the number of observations.

v

Numeric for resolution of data or a vector specifying the observation points.

rho

Numeric which indicates the dependence on the previous curve.

Value

A dfts object for the generated data.

Examples

generate_ornstein_uhlenbeck(N=100,v=20)
#> data 
#> dimension: 20 x 100
#> fparam (20): 0 0.05263158 0.1052632 0.1578947 0.2105263 0.2631579 ... 
#> labels (100): X1 X2 X3 X4 X5 X6 ... 
#> 
#> Example data
#>             [,1]      [,2]       [,3]       [,4]       [,5]      [,6]
#> [1,] -0.29620749 1.4498754 -0.4510367 -0.9636357  0.1805783 0.7330501
#> [2,] -0.07482246 0.9855025 -0.5535832 -0.6943754 -0.1613623 0.8615638
#> [3,] -0.15176517 0.5976816 -0.6665618 -1.1725090 -0.1291813 0.8853528
#> [4,] -0.00672887 0.2356879 -0.8687792 -1.2242976 -0.1772623 0.8758123
#> [5,] -0.45065040 0.4904342 -0.9817375 -1.1326045 -0.1867702 0.7531731
#> [6,]  0.01121310 0.2072418 -1.0923997 -0.6848316 -0.3262242 0.6543656