Functional KPSS Test

Compute the KPSS statistic for functional data.

Usage

kpss_test(
  X,
  method = c("MC", "block", "sliding"),
  M = 1000,
  h = 3,
  TVE = 1,
  replace = TRUE
)

Arguments

X

A dfts object or data which can be automatically converted to that format. See dfts().

method

String (default MC) for the method: monte carlo simulation (MC), block permutation (block), and sliding window permutation (sliding)

M

Number of simulations (default 1000) to estimate theoretical distribution

h

Numeric (default 3) for the block size when using a permutation test

TVE

Numeric (default 1) for selecting the number of principle components / eigenvalues

replace

Boolean (default TRUE) to indicate if blocks should be selected with replacement when using a permutation test

Value

List with statistic, p-value and theoretical simulations, respectively

References

Chen, Y., & Pun, C. S. (2019). A bootstrap-based KPSS test for functional time series. Journal of Multivariate Analysis, 174, 104535.

Kokoszka, P., & Young, G. (2016). KPSS test for functional time series. Statistics, 50(5), 957-973.

Examples

kpss_test(generate_brownian_motion(100,v=seq(0,1,length.out=20)))
#> $statistic
#> [1] 0.02003026
#> 
#> $pvalue
#> [1] 0.858
#> 
#> $sims
#>    [1] 0.030235788 0.031324327 0.060389517 0.014029960 0.045126232 0.041404468
#>    [7] 0.029629899 0.033970911 0.033600012 0.031522996 0.104084602 0.042139261
#>   [13] 0.064147236 0.038944328 0.016627924 0.056764070 0.034380121 0.022808918
#>   [19] 0.031191548 0.023952323 0.047011466 0.013958716 0.037369998 0.016282510
#>   [25] 0.020812269 0.013662363 0.034919122 0.032619809 0.148344074 0.047967713
#>   [31] 0.025388206 0.031574596 0.062814164 0.011834259 0.016986772 0.038255664
#>   [37] 0.106678804 0.020542299 0.030877620 0.044230778 0.065122720 0.040520308
#>   [43] 0.027115000 0.064135207 0.027597433 0.052943053 0.021831219 0.085807946
#>   [49] 0.038137703 0.040165560 0.029871138 0.035625778 0.037093315 0.040395225
#>   [55] 0.037135287 0.018182122 0.020500661 0.037520460 0.064582529 0.018224555
#>   [61] 0.020979402 0.027591062 0.032549666 0.026664270 0.059256462 0.034689528
#>   [67] 0.073183007 0.022446521 0.016381475 0.017368872 0.024742659 0.030172079
#>   [73] 0.061541898 0.031809222 0.033222076 0.031471233 0.032501084 0.025515232
#>   [79] 0.021802924 0.044299256 0.035957604 0.030946063 0.031158218 0.014827369
#>   [85] 0.026068966 0.045906833 0.013859239 0.041962242 0.022848892 0.017151509
#>   [91] 0.033477669 0.022894503 0.113084180 0.016750591 0.018952394 0.023574846
#>   [97] 0.029479864 0.044782352 0.024710046 0.023112938 0.029927271 0.023990905
#>  [103] 0.050488034 0.033263409 0.118091960 0.017423583 0.013228398 0.047019048
#>  [109] 0.020278755 0.038682734 0.031619643 0.029306003 0.030150807 0.037619685
#>  [115] 0.048667252 0.017020130 0.021080800 0.034325029 0.029975792 0.056085354
#>  [121] 0.072629648 0.032006397 0.027440006 0.016501330 0.008410735 0.046685208
#>  [127] 0.024366612 0.040712947 0.055992732 0.034826694 0.042180386 0.040525487
#>  [133] 0.032015949 0.022597438 0.042773863 0.052395718 0.028114604 0.056041341
#>  [139] 0.036061882 0.051095563 0.050557858 0.059104189 0.040474811 0.032238039
#>  [145] 0.054308755 0.019062100 0.066333536 0.033809536 0.040631787 0.017951271
#>  [151] 0.035920624 0.024562502 0.036800462 0.023857032 0.043475869 0.025131155
#>  [157] 0.019226170 0.017314247 0.078886521 0.029431362 0.017057241 0.025494824
#>  [163] 0.043548535 0.019550385 0.021149480 0.012001254 0.031446771 0.029431672
#>  [169] 0.016683740 0.023868667 0.013885874 0.047642612 0.033937242 0.043436019
#>  [175] 0.034414840 0.040814157 0.023598981 0.072165948 0.021860750 0.027613833
#>  [181] 0.026083127 0.038360129 0.054023749 0.038137142 0.033058877 0.025308006
#>  [187] 0.010485368 0.069015592 0.019836725 0.015216234 0.020995461 0.028511559
#>  [193] 0.031733597 0.075241745 0.044583194 0.053060207 0.035535329 0.051432826
#>  [199] 0.071285881 0.043578033 0.039089512 0.026669190 0.142345737 0.033205908
#>  [205] 0.017373381 0.019822920 0.018483132 0.032438078 0.051436390 0.026988466
#>  [211] 0.023491466 0.052010693 0.039045804 0.025539245 0.023195909 0.034138210
#>  [217] 0.018225866 0.017359876 0.013117182 0.046029851 0.030770881 0.044118370
#>  [223] 0.044270460 0.026421775 0.057515644 0.019577420 0.023796308 0.038214531
#>  [229] 0.024631305 0.046306141 0.027388967 0.037737291 0.086708807 0.047614194
#>  [235] 0.031150416 0.074048442 0.060676983 0.073332030 0.020739218 0.033261134
#>  [241] 0.021580839 0.061594548 0.022147967 0.019812271 0.025331905 0.031040804
#>  [247] 0.042016448 0.020594332 0.021106996 0.026692587 0.038072420 0.062604793
#>  [253] 0.022731228 0.033873591 0.018692526 0.044722167 0.027905064 0.021662972
#>  [259] 0.032999071 0.019702297 0.065486661 0.031538347 0.013416629 0.057980226
#>  [265] 0.016872770 0.018977909 0.010474229 0.041703908 0.020483883 0.112160117
#>  [271] 0.035135724 0.022171216 0.020954020 0.062079738 0.023714439 0.056970558
#>  [277] 0.017585549 0.016150902 0.016943175 0.054202428 0.018736734 0.027224523
#>  [283] 0.042099134 0.033145000 0.026381492 0.037996126 0.023115764 0.025341476
#>  [289] 0.020507129 0.022541674 0.054321113 0.043035538 0.049399931 0.022047612
#>  [295] 0.066464264 0.024845026 0.019993062 0.048292003 0.091865845 0.049084163
#>  [301] 0.027547247 0.025023398 0.038984487 0.017861925 0.053591708 0.034906651
#>  [307] 0.033205745 0.024246755 0.036978019 0.060400463 0.017981047 0.060230946
#>  [313] 0.050759578 0.061853270 0.072352050 0.040310941 0.054408596 0.118801513
#>  [319] 0.031463302 0.022517205 0.029371650 0.015491262 0.030594213 0.030765461
#>  [325] 0.022048429 0.026959714 0.035730466 0.018006626 0.035675155 0.027346869
#>  [331] 0.049306927 0.074341992 0.037810437 0.031104545 0.025540979 0.028172455
#>  [337] 0.013987415 0.023957440 0.044967079 0.058489944 0.048011067 0.048289026
#>  [343] 0.032671483 0.013476976 0.073125310 0.020964821 0.042824658 0.045678517
#>  [349] 0.024718550 0.022588629 0.023778277 0.039237948 0.074287458 0.063293797
#>  [355] 0.027853929 0.024483398 0.020661775 0.021516338 0.031004872 0.115178487
#>  [361] 0.032028321 0.028549241 0.030308852 0.055094194 0.024170865 0.047235056
#>  [367] 0.034271731 0.021930392 0.021599891 0.033142817 0.034328866 0.072936104
#>  [373] 0.021378297 0.020252149 0.057994092 0.014406534 0.056464548 0.036180386
#>  [379] 0.028401764 0.055321336 0.023848210 0.021144884 0.030454902 0.020528053
#>  [385] 0.034875523 0.031220690 0.055029972 0.014052481 0.043009987 0.027681567
#>  [391] 0.027647503 0.019299412 0.045075353 0.049923346 0.070665585 0.031646569
#>  [397] 0.024639086 0.027169359 0.042479344 0.062390510 0.021387998 0.015351039
#>  [403] 0.048718003 0.058787015 0.042045186 0.021597091 0.040974455 0.029388808
#>  [409] 0.051585228 0.054620121 0.086762095 0.058394247 0.047351518 0.033432842
#>  [415] 0.047386920 0.024473735 0.026951891 0.021029566 0.052405033 0.032467821
#>  [421] 0.015689377 0.035216566 0.032143067 0.043085445 0.051623454 0.045221421
#>  [427] 0.026437925 0.027334323 0.028663011 0.021981093 0.052864706 0.021638505
#>  [433] 0.048767757 0.046806934 0.029703962 0.029776089 0.017751112 0.026047353
#>  [439] 0.080546631 0.048333127 0.023457307 0.037039177 0.038812151 0.038863352
#>  [445] 0.064855326 0.039402776 0.047585844 0.043565480 0.038831876 0.036071866
#>  [451] 0.030574049 0.023186867 0.028909164 0.021825861 0.042192576 0.037988052
#>  [457] 0.028950364 0.025659617 0.042787043 0.042431619 0.028069537 0.049307723
#>  [463] 0.035467983 0.036282376 0.021610366 0.022834931 0.024033341 0.175715296
#>  [469] 0.027404461 0.025321260 0.027285692 0.049024582 0.028411688 0.039472965
#>  [475] 0.023207296 0.027746592 0.048489777 0.025440123 0.052478876 0.015777164
#>  [481] 0.053115423 0.044993992 0.033301317 0.037822510 0.073804955 0.033207728
#>  [487] 0.050302076 0.012685155 0.024227018 0.022254397 0.054071988 0.035601465
#>  [493] 0.046591716 0.023834528 0.062882053 0.020608909 0.028535576 0.120105435
#>  [499] 0.028314992 0.021964479 0.036458622 0.026265790 0.036263809 0.038750614
#>  [505] 0.041286064 0.021810018 0.038860268 0.044308609 0.026630491 0.018069382
#>  [511] 0.043114407 0.016229351 0.066304703 0.020063848 0.017270398 0.042725581
#>  [517] 0.142839986 0.043752028 0.028966598 0.032802298 0.063697301 0.021845475
#>  [523] 0.016320518 0.036822085 0.024381837 0.035547371 0.030337833 0.018998062
#>  [529] 0.035114705 0.027379888 0.026256771 0.030752762 0.025483332 0.038306073
#>  [535] 0.029400461 0.025081229 0.021209394 0.021948305 0.036086554 0.038955185
#>  [541] 0.024924966 0.057496970 0.046258542 0.040537010 0.022850728 0.030435591
#>  [547] 0.029682070 0.039967122 0.030370707 0.049479192 0.015748013 0.058105796
#>  [553] 0.032903797 0.030985134 0.053428374 0.034281726 0.163733006 0.019759844
#>  [559] 0.037039149 0.044425525 0.037593142 0.028925790 0.032166426 0.051164598
#>  [565] 0.044107338 0.022171905 0.076674683 0.028764687 0.024229962 0.078960515
#>  [571] 0.037620332 0.049942803 0.069781358 0.042198235 0.044758727 0.024823962
#>  [577] 0.022726700 0.031658817 0.036358769 0.056984268 0.040273729 0.027180881
#>  [583] 0.023602570 0.053426131 0.030087559 0.048355283 0.017011383 0.044156244
#>  [589] 0.026407754 0.023785566 0.039888656 0.034469016 0.016388642 0.038561019
#>  [595] 0.038842830 0.016767150 0.014318129 0.043706738 0.031081400 0.031789208
#>  [601] 0.064787097 0.021678150 0.035875841 0.022109303 0.062890306 0.038098742
#>  [607] 0.038737132 0.052724630 0.056502701 0.016017947 0.088126439 0.032341911
#>  [613] 0.031455773 0.014453655 0.028727136 0.024095633 0.036500330 0.020061130
#>  [619] 0.038216530 0.019006546 0.023179436 0.027644170 0.020887732 0.018789237
#>  [625] 0.029073277 0.035547066 0.012262529 0.031537348 0.024913987 0.057394752
#>  [631] 0.037708540 0.029044614 0.048405803 0.035874832 0.050643419 0.015614725
#>  [637] 0.022268259 0.042000817 0.113429943 0.041016023 0.029441898 0.011434304
#>  [643] 0.022839015 0.030830160 0.022138181 0.033103855 0.014315879 0.025102152
#>  [649] 0.020107838 0.013306593 0.020921655 0.015784133 0.034305401 0.054536850
#>  [655] 0.041439174 0.036885863 0.045180616 0.071434780 0.048001797 0.027066108
#>  [661] 0.016329514 0.023034543 0.033987315 0.029488553 0.050588609 0.075263795
#>  [667] 0.048366098 0.016830586 0.032998640 0.034451542 0.014379472 0.020836074
#>  [673] 0.037569121 0.018266088 0.035579802 0.031742537 0.031411267 0.068164829
#>  [679] 0.022051953 0.039260576 0.010175359 0.025272235 0.049376583 0.012832941
#>  [685] 0.018496300 0.026842107 0.044138361 0.041410013 0.039251823 0.078302275
#>  [691] 0.013135441 0.023832642 0.031125251 0.022824028 0.039077888 0.028133216
#>  [697] 0.020237166 0.025037233 0.045008226 0.049612367 0.069782847 0.014240538
#>  [703] 0.060907327 0.014416590 0.080620940 0.023418974 0.054638417 0.032237787
#>  [709] 0.027136016 0.038256219 0.033659029 0.024396137 0.018597628 0.021518407
#>  [715] 0.026071648 0.026373922 0.029117752 0.057563932 0.038479795 0.025248088
#>  [721] 0.017581625 0.050250321 0.072722134 0.038454831 0.042270113 0.034334537
#>  [727] 0.038548840 0.036256518 0.025990122 0.017825398 0.038254257 0.075748968
#>  [733] 0.036117329 0.023153408 0.057973552 0.038741521 0.060892841 0.051524650
#>  [739] 0.033115043 0.021584717 0.035560523 0.022223203 0.023702926 0.059332089
#>  [745] 0.041803895 0.070032289 0.026140414 0.023510659 0.038495042 0.022148663
#>  [751] 0.038645590 0.024743114 0.054251475 0.047465846 0.011233931 0.087413098
#>  [757] 0.031549303 0.015913231 0.033612687 0.018090272 0.057019651 0.020140956
#>  [763] 0.025300215 0.023339586 0.011401164 0.026551335 0.056987607 0.045317745
#>  [769] 0.028176520 0.019726903 0.034138128 0.034270288 0.025604310 0.016741610
#>  [775] 0.028526966 0.035290390 0.025912579 0.032320333 0.054658591 0.026856645
#>  [781] 0.025265757 0.028174913 0.024278437 0.026895148 0.042653410 0.020642970
#>  [787] 0.022677132 0.048794678 0.031331450 0.041418061 0.015390130 0.015185669
#>  [793] 0.016325349 0.021053772 0.021750893 0.016980683 0.029089765 0.097492573
#>  [799] 0.032178855 0.069025737 0.027328657 0.022155957 0.034380422 0.057075581
#>  [805] 0.049123545 0.056220318 0.027712186 0.057615753 0.035979980 0.029956569
#>  [811] 0.047709581 0.021770104 0.028772915 0.036187863 0.013730632 0.015682558
#>  [817] 0.041242585 0.033454722 0.017500575 0.017610402 0.049990172 0.051564840
#>  [823] 0.024717476 0.077918351 0.044400451 0.067617945 0.010121580 0.017767196
#>  [829] 0.035558865 0.039444467 0.097865508 0.027622284 0.073709016 0.029086149
#>  [835] 0.038929623 0.042938471 0.030514459 0.039911501 0.024266345 0.049619774
#>  [841] 0.050921398 0.024731729 0.046611668 0.053249364 0.008991313 0.019486980
#>  [847] 0.019003923 0.033737712 0.022935235 0.026010724 0.150267855 0.032779774
#>  [853] 0.024229718 0.044612102 0.040775244 0.052476353 0.016834031 0.019301995
#>  [859] 0.031143021 0.033515751 0.097572404 0.025210441 0.020722020 0.017369281
#>  [865] 0.021556116 0.024593538 0.046658793 0.034468196 0.027400147 0.048938171
#>  [871] 0.027459072 0.051879058 0.023322313 0.019904183 0.038657824 0.031977572
#>  [877] 0.020018085 0.053059814 0.041459333 0.022221409 0.038285878 0.022851522
#>  [883] 0.044128833 0.031102961 0.020558841 0.043563423 0.037021661 0.032498170
#>  [889] 0.023544437 0.058971331 0.030029439 0.037628363 0.015279025 0.090678190
#>  [895] 0.019315232 0.024625470 0.026219995 0.018699038 0.025796936 0.020224512
#>  [901] 0.015329562 0.024341331 0.036801079 0.025423834 0.069227609 0.034355185
#>  [907] 0.033300910 0.026829660 0.108525186 0.023279305 0.021389870 0.045911566
#>  [913] 0.074314937 0.044221553 0.018839779 0.043720420 0.026575841 0.037815948
#>  [919] 0.040807908 0.024794416 0.039657610 0.016144410 0.022257736 0.023215878
#>  [925] 0.033339155 0.048829692 0.019237986 0.017989607 0.036808399 0.021395917
#>  [931] 0.050391000 0.050535824 0.021542467 0.049123211 0.042488616 0.028532086
#>  [937] 0.022060011 0.033365409 0.021844918 0.013341856 0.008320795 0.026466777
#>  [943] 0.118048051 0.023112595 0.019984823 0.050552956 0.040643648 0.025193065
#>  [949] 0.020066164 0.012583949 0.021176170 0.037999699 0.027406085 0.028748643
#>  [955] 0.015782770 0.013266923 0.041102292 0.035367247 0.024883182 0.033682508
#>  [961] 0.018339049 0.042699736 0.029130174 0.034883904 0.045396577 0.036260293
#>  [967] 0.024506754 0.044413793 0.026485668 0.025775209 0.076503017 0.045768249
#>  [973] 0.031923454 0.032373872 0.046386320 0.057555735 0.037836088 0.062509425
#>  [979] 0.032007293 0.178079655 0.059737642 0.019322850 0.020968246 0.051170168
#>  [985] 0.047187710 0.023971346 0.061515523 0.035592824 0.026924454 0.042416500
#>  [991] 0.010561794 0.046104146 0.038443425 0.031200017 0.038593989 0.040785979
#>  [997] 0.018051686 0.072760585 0.050667324 0.030015604
#> 
kpss_test(generate_brownian_motion(100,v=seq(0,1,length.out=20)),
             method="block")
#> $statistic
#> [1] 0.02017272
#> 
#> $pvalue
#> [1] 0.846
#> 
#> $sims
#>    [1] 0.036726078 0.024450331 0.012016947 0.029062918 0.017218399 0.027661883
#>    [7] 0.034284910 0.039475189 0.038089904 0.051462093 0.066878989 0.050118899
#>   [13] 0.064226770 0.048875079 0.089081004 0.023746933 0.056181582 0.030320147
#>   [19] 0.033251719 0.064907919 0.037852097 0.024111059 0.026637572 0.022150467
#>   [25] 0.059663063 0.013931756 0.065446920 0.035487639 0.034871170 0.021851051
#>   [31] 0.052287719 0.066425541 0.028143646 0.025324451 0.020233673 0.017252561
#>   [37] 0.030332962 0.058404289 0.026195357 0.144300635 0.018560083 0.019065631
#>   [43] 0.036028068 0.040455415 0.046259887 0.046217815 0.019132158 0.020794473
#>   [49] 0.015790053 0.051493600 0.069813487 0.028688845 0.029762786 0.019832185
#>   [55] 0.038488736 0.023147088 0.020864074 0.047442731 0.039363452 0.019431427
#>   [61] 0.018755195 0.032038301 0.043616090 0.030580539 0.024229386 0.062158013
#>   [67] 0.023755792 0.040128293 0.019546161 0.023408663 0.019156622 0.030704144
#>   [73] 0.043416639 0.018609087 0.058860618 0.036551396 0.048724205 0.028307733
#>   [79] 0.060298428 0.031443409 0.033408436 0.021368086 0.046941571 0.018584942
#>   [85] 0.050358131 0.076133068 0.027363850 0.021664588 0.024020126 0.023496229
#>   [91] 0.056574084 0.030462424 0.025264134 0.038224937 0.016085592 0.029299696
#>   [97] 0.024691257 0.018452886 0.083981179 0.014534872 0.020537570 0.025505307
#>  [103] 0.018225309 0.030724538 0.017387090 0.033036701 0.064280047 0.029324179
#>  [109] 0.031260209 0.019616320 0.061978504 0.020201299 0.022219513 0.031331357
#>  [115] 0.020726180 0.058437848 0.031203020 0.023826630 0.030236214 0.047288857
#>  [121] 0.032359976 0.018857520 0.019014478 0.034317663 0.017352547 0.028423722
#>  [127] 0.017157558 0.030945248 0.021262086 0.051124120 0.043785368 0.024334594
#>  [133] 0.031578149 0.050393375 0.023995483 0.011240329 0.043245295 0.036409605
#>  [139] 0.041250849 0.042280510 0.038630053 0.020535112 0.043500738 0.019531439
#>  [145] 0.021051693 0.034493378 0.020670410 0.029978334 0.055691566 0.020382851
#>  [151] 0.025753893 0.043057476 0.021166486 0.041567307 0.019154493 0.035779424
#>  [157] 0.023266550 0.034586548 0.034293783 0.033885941 0.018605873 0.014445748
#>  [163] 0.043652778 0.048476005 0.065120655 0.021641585 0.026558793 0.032039697
#>  [169] 0.037848263 0.020899345 0.031394559 0.027615201 0.070015608 0.062170597
#>  [175] 0.020574694 0.035207569 0.037584030 0.055123550 0.049413375 0.046248306
#>  [181] 0.027731802 0.034851376 0.020562128 0.021503968 0.048464342 0.043585403
#>  [187] 0.026088815 0.048780183 0.033279379 0.025852227 0.016407939 0.019321471
#>  [193] 0.016391570 0.022497997 0.038342503 0.057038502 0.045326396 0.048667785
#>  [199] 0.044352497 0.030887620 0.022356847 0.082316942 0.047010331 0.018634040
#>  [205] 0.015920317 0.033328494 0.047815661 0.023596673 0.052662692 0.058475412
#>  [211] 0.024042023 0.025090033 0.069026852 0.050948961 0.022079795 0.033884620
#>  [217] 0.018545330 0.026101653 0.034491908 0.026247709 0.020948667 0.018769373
#>  [223] 0.034266698 0.042993302 0.047079372 0.017908331 0.033573321 0.023362412
#>  [229] 0.028649620 0.044415852 0.032844354 0.050552370 0.030801892 0.027310327
#>  [235] 0.067249082 0.036412940 0.045411994 0.031938383 0.072390818 0.037181495
#>  [241] 0.033064159 0.016880682 0.039179131 0.066778801 0.064380473 0.045915147
#>  [247] 0.021417283 0.024242985 0.015790776 0.023804837 0.017207519 0.018872584
#>  [253] 0.028160075 0.016179618 0.080374545 0.016753352 0.016848008 0.055036180
#>  [259] 0.036169402 0.031516906 0.024164143 0.047832973 0.059369058 0.023286752
#>  [265] 0.058874216 0.021118907 0.019641329 0.022087517 0.041938124 0.049119496
#>  [271] 0.021367785 0.037273389 0.048546554 0.018822776 0.077158197 0.024839113
#>  [277] 0.028041162 0.049798477 0.018191535 0.027316919 0.041498023 0.018523959
#>  [283] 0.064070142 0.039929117 0.037538763 0.018268679 0.018412816 0.023735190
#>  [289] 0.014958127 0.014798090 0.026006373 0.046902644 0.047036059 0.053829510
#>  [295] 0.048539377 0.020243881 0.062915373 0.042827151 0.053995704 0.044297562
#>  [301] 0.024549027 0.033747868 0.016810320 0.053914146 0.028393839 0.039732495
#>  [307] 0.027141344 0.063915887 0.034699056 0.033829032 0.030317216 0.026620572
#>  [313] 0.031743007 0.024272607 0.053485934 0.076982969 0.062848270 0.033585751
#>  [319] 0.031157129 0.042814563 0.031871882 0.079784286 0.021907147 0.029903660
#>  [325] 0.040511603 0.067640003 0.029566828 0.029548847 0.024599675 0.021903868
#>  [331] 0.029211949 0.027902787 0.036181086 0.040129852 0.012106316 0.073418469
#>  [337] 0.021358762 0.031585089 0.031161317 0.018515421 0.061040260 0.037296638
#>  [343] 0.052004930 0.027871848 0.041961801 0.034868706 0.022253338 0.034367668
#>  [349] 0.019031332 0.037873318 0.015891567 0.027654577 0.013406660 0.030435586
#>  [355] 0.031159610 0.061676199 0.021782210 0.022396228 0.045720501 0.015083948
#>  [361] 0.038777758 0.023819742 0.014061396 0.036372834 0.046392482 0.015364604
#>  [367] 0.042022744 0.025451558 0.025669900 0.023970252 0.086072479 0.016224219
#>  [373] 0.026883993 0.021785955 0.045222240 0.039250372 0.019551953 0.027001224
#>  [379] 0.034393521 0.026159916 0.013709097 0.023822488 0.025866021 0.020570923
#>  [385] 0.105915040 0.019392925 0.025338180 0.043787457 0.011758473 0.053408356
#>  [391] 0.069577832 0.018872875 0.066064017 0.027559302 0.043589832 0.045780682
#>  [397] 0.040324017 0.029838381 0.018439390 0.025714918 0.022839915 0.028498423
#>  [403] 0.027158362 0.052935794 0.028589982 0.112791011 0.040432849 0.025669298
#>  [409] 0.026823485 0.053980255 0.035230005 0.024724084 0.045732110 0.068474633
#>  [415] 0.026532999 0.063286358 0.025267623 0.021777340 0.061466232 0.034296678
#>  [421] 0.078069181 0.083610650 0.056530962 0.029986465 0.029860528 0.042949073
#>  [427] 0.051123161 0.026384548 0.027630500 0.030249215 0.036342576 0.024716282
#>  [433] 0.026534175 0.034955098 0.062139895 0.043334910 0.018574128 0.024587451
#>  [439] 0.075564222 0.024004043 0.087873813 0.018023099 0.038785101 0.091443450
#>  [445] 0.017083594 0.022750154 0.029639968 0.025930159 0.038277861 0.050669287
#>  [451] 0.038015562 0.067244040 0.027987941 0.032560258 0.025028421 0.033280352
#>  [457] 0.020221648 0.021735935 0.068335027 0.054456261 0.022656920 0.033444927
#>  [463] 0.079127663 0.034685408 0.034444236 0.024871260 0.016569634 0.045572595
#>  [469] 0.026029277 0.021613486 0.044536261 0.036298529 0.067636208 0.015578806
#>  [475] 0.022848173 0.013359805 0.076900015 0.033680055 0.038625757 0.015668978
#>  [481] 0.019553223 0.018012227 0.098156745 0.016908107 0.026138742 0.046972081
#>  [487] 0.032239598 0.028007212 0.024731570 0.019905180 0.087337579 0.044942911
#>  [493] 0.030685059 0.062783145 0.054269303 0.010919679 0.018074231 0.043789439
#>  [499] 0.020484063 0.020514346 0.042488903 0.027726070 0.028435694 0.023681014
#>  [505] 0.045613343 0.018227970 0.050396694 0.029410111 0.016825146 0.042211670
#>  [511] 0.031046054 0.040238924 0.029530317 0.037910742 0.030124115 0.020743668
#>  [517] 0.082104511 0.025412073 0.046501758 0.026626870 0.030272412 0.042475487
#>  [523] 0.020299958 0.024748386 0.011341215 0.044490693 0.028376976 0.060367106
#>  [529] 0.018265727 0.031367317 0.050049734 0.015511102 0.022967588 0.039841156
#>  [535] 0.018256282 0.018587349 0.042837660 0.065621771 0.030723581 0.018298695
#>  [541] 0.023707528 0.071569655 0.130761291 0.064201881 0.067030922 0.030637337
#>  [547] 0.039354820 0.037440500 0.021771411 0.041581203 0.040266614 0.025692949
#>  [553] 0.033073387 0.043718547 0.026096980 0.024406892 0.061589359 0.032898159
#>  [559] 0.044895740 0.040054002 0.031590380 0.042730246 0.033372944 0.028574741
#>  [565] 0.039652985 0.028773372 0.057134409 0.042206352 0.026288867 0.055697673
#>  [571] 0.074260882 0.016362937 0.034059293 0.040001885 0.028745133 0.020095467
#>  [577] 0.026159053 0.021988293 0.034311730 0.032296444 0.025278288 0.054975325
#>  [583] 0.042614038 0.023448943 0.033653584 0.019189543 0.062333915 0.041388015
#>  [589] 0.057569210 0.024577195 0.020588031 0.027617930 0.028529308 0.043174849
#>  [595] 0.027426339 0.025898589 0.029123739 0.028146195 0.021754448 0.031349721
#>  [601] 0.042879764 0.049540833 0.025214861 0.023294548 0.022234249 0.021029547
#>  [607] 0.023853470 0.039919814 0.032083929 0.020647179 0.031878486 0.052527493
#>  [613] 0.058646473 0.021741184 0.061054044 0.019918825 0.032077677 0.017640754
#>  [619] 0.035465359 0.056402044 0.019278235 0.026863781 0.023522658 0.037446164
#>  [625] 0.026677020 0.029744462 0.031534997 0.027296119 0.035138275 0.016252586
#>  [631] 0.022346547 0.053629099 0.023844113 0.043140661 0.026751664 0.032184932
#>  [637] 0.028862822 0.021778127 0.023507669 0.062116112 0.104202609 0.026834136
#>  [643] 0.015836638 0.017538717 0.036421822 0.015748091 0.031224440 0.028382179
#>  [649] 0.034404120 0.031511579 0.022365291 0.024439630 0.024484344 0.068389822
#>  [655] 0.026375399 0.046139292 0.040592708 0.029163293 0.042993805 0.029191125
#>  [661] 0.013915994 0.023955023 0.029842839 0.027958283 0.062004494 0.025225849
#>  [667] 0.031977454 0.029417640 0.040366407 0.055609643 0.016037631 0.022785343
#>  [673] 0.035395731 0.029166142 0.018942566 0.046792047 0.017336232 0.032523507
#>  [679] 0.076831707 0.056115605 0.011823939 0.031561607 0.024802190 0.047323060
#>  [685] 0.027158160 0.020574347 0.018929858 0.063154151 0.022121518 0.079588536
#>  [691] 0.016493494 0.019408843 0.050409870 0.027738456 0.028676270 0.051056319
#>  [697] 0.034373021 0.016557580 0.046419898 0.027178790 0.030220090 0.025842365
#>  [703] 0.030309658 0.050414115 0.092511026 0.074596067 0.021892669 0.038050835
#>  [709] 0.038543820 0.021482295 0.036712427 0.043154066 0.010892422 0.018498191
#>  [715] 0.066451643 0.034133635 0.027984933 0.026292124 0.015192926 0.035100479
#>  [721] 0.028895213 0.052179612 0.045104924 0.024615896 0.021583769 0.025654364
#>  [727] 0.028700781 0.012390023 0.025093000 0.042868017 0.027449092 0.022898957
#>  [733] 0.018351343 0.042544782 0.104111758 0.031419015 0.078826815 0.021198853
#>  [739] 0.045213034 0.013715578 0.019669981 0.038285908 0.034793359 0.035317254
#>  [745] 0.019645552 0.015052917 0.013202627 0.032510430 0.026600586 0.035723528
#>  [751] 0.024192585 0.026883638 0.035471707 0.029982185 0.023704076 0.028588439
#>  [757] 0.051548327 0.052228402 0.057425325 0.045505007 0.046092504 0.034584002
#>  [763] 0.026305559 0.045411674 0.034542439 0.047291144 0.020956707 0.041053205
#>  [769] 0.043625416 0.023568047 0.036153367 0.034662424 0.019738629 0.040519060
#>  [775] 0.029392859 0.052805791 0.028385147 0.043562048 0.055847969 0.029363674
#>  [781] 0.011886941 0.034626260 0.104154481 0.026519684 0.043506817 0.023118955
#>  [787] 0.015859995 0.035352993 0.020989108 0.062219949 0.043933425 0.026035096
#>  [793] 0.022722105 0.017958400 0.065196936 0.058831744 0.018843487 0.056880219
#>  [799] 0.033472286 0.026509272 0.039739459 0.026713759 0.017374599 0.024122616
#>  [805] 0.032399360 0.018746142 0.033120924 0.033743868 0.014994213 0.058615238
#>  [811] 0.034608388 0.047865012 0.036257274 0.044651603 0.042117213 0.036890037
#>  [817] 0.054347617 0.024999397 0.054571353 0.013572071 0.033623960 0.036667986
#>  [823] 0.091602016 0.059289547 0.032838141 0.022838178 0.027846491 0.026827611
#>  [829] 0.062197498 0.028653188 0.030992890 0.068527091 0.019354476 0.052982193
#>  [835] 0.031481821 0.033081078 0.015547150 0.103202333 0.050041380 0.064216874
#>  [841] 0.022967271 0.029128482 0.018937820 0.027246527 0.021947493 0.033575426
#>  [847] 0.016385957 0.019311245 0.028441313 0.024329313 0.065243041 0.055376113
#>  [853] 0.044421445 0.018809119 0.047710386 0.022667661 0.037679045 0.017392655
#>  [859] 0.042814711 0.032835235 0.030526654 0.068660558 0.021260035 0.036616003
#>  [865] 0.046820977 0.031528979 0.030036506 0.072228612 0.020956153 0.081301989
#>  [871] 0.045580994 0.009024575 0.051602158 0.020272277 0.020287991 0.089940812
#>  [877] 0.120027702 0.038677131 0.050334846 0.014779361 0.029067946 0.018563976
#>  [883] 0.035563035 0.023482076 0.022299345 0.031858529 0.031407258 0.028967047
#>  [889] 0.098822906 0.036447752 0.011558809 0.026274097 0.030196681 0.039166835
#>  [895] 0.042141034 0.033663957 0.028627243 0.023065938 0.043340161 0.031692921
#>  [901] 0.031994088 0.037127985 0.043077868 0.030431564 0.037255996 0.025808502
#>  [907] 0.027516101 0.038095888 0.038177825 0.063605577 0.041118920 0.054084523
#>  [913] 0.028190846 0.019837227 0.032870486 0.023989324 0.020680004 0.028006971
#>  [919] 0.027652030 0.020710483 0.027042615 0.045120547 0.026332625 0.028599833
#>  [925] 0.032865658 0.039211108 0.037836712 0.012636021 0.019866370 0.055628774
#>  [931] 0.017890453 0.048720995 0.020122738 0.019800341 0.026797408 0.018453974
#>  [937] 0.030340564 0.091928246 0.049735056 0.030237061 0.043472005 0.080750272
#>  [943] 0.016991402 0.026695981 0.049318048 0.032037969 0.032365411 0.017765440
#>  [949] 0.055731633 0.019388544 0.035860230 0.073467675 0.026799473 0.013214700
#>  [955] 0.061492824 0.038948726 0.044357302 0.024975576 0.039820120 0.032196961
#>  [961] 0.064987254 0.020905329 0.043586437 0.021762261 0.029273097 0.019448945
#>  [967] 0.030904252 0.026333666 0.042269574 0.033410465 0.046589745 0.033190249
#>  [973] 0.039846048 0.031751077 0.034414937 0.032403789 0.021852404 0.128707799
#>  [979] 0.032270584 0.021356912 0.031714778 0.016690703 0.037457130 0.017632422
#>  [985] 0.024435173 0.083761539 0.045303620 0.023929400 0.025789662 0.038820633
#>  [991] 0.032761599 0.022143430 0.034829172 0.040210455 0.028660986 0.035378282
#>  [997] 0.020828311 0.017669457 0.036167681 0.046499825
#> 
kpss_test(generate_brownian_motion(100,v=seq(0,1,length.out=20)),
             method='sliding')
#> $statistic
#> [1] 0.05354905
#> 
#> $pvalue
#> [1] 0.126
#> 
#> $sims
#>    [1] 0.024454606 0.012336282 0.024060386 0.031167795 0.029968020 0.024677082
#>    [7] 0.099431269 0.011328104 0.020880780 0.042621117 0.029274591 0.055393992
#>   [13] 0.013993012 0.057491770 0.051194928 0.054191583 0.037408474 0.013720658
#>   [19] 0.012741220 0.043566483 0.015845459 0.034113968 0.028380049 0.017742792
#>   [25] 0.054117597 0.022060574 0.019152016 0.043798507 0.062946013 0.048762618
#>   [31] 0.048661450 0.055990908 0.028282185 0.050596607 0.022285520 0.015432785
#>   [37] 0.026029082 0.021144106 0.031740686 0.016892286 0.045361185 0.027300684
#>   [43] 0.075191833 0.015889649 0.063879063 0.021020905 0.047121756 0.067836299
#>   [49] 0.023631112 0.030607465 0.031224858 0.027357646 0.016943868 0.014644298
#>   [55] 0.036302715 0.029850779 0.055328383 0.032332503 0.105229732 0.032991675
#>   [61] 0.030156247 0.023509903 0.011644295 0.036931226 0.037050695 0.047639322
#>   [67] 0.048742563 0.014393747 0.073220059 0.048329939 0.032743943 0.042316266
#>   [73] 0.025384195 0.027495705 0.060904130 0.046079421 0.013502988 0.024592697
#>   [79] 0.029493562 0.032954068 0.099898479 0.017189464 0.046841259 0.013776641
#>   [85] 0.013785316 0.031963179 0.020494166 0.070300207 0.038218639 0.042232832
#>   [91] 0.075205101 0.025086197 0.027211616 0.024923846 0.046180244 0.063975270
#>   [97] 0.024550062 0.013519094 0.056715897 0.037434513 0.045948721 0.047070398
#>  [103] 0.028589127 0.025377961 0.051258162 0.020345696 0.049684750 0.044229299
#>  [109] 0.027116166 0.017244413 0.016552877 0.024879449 0.033438849 0.023486572
#>  [115] 0.013709984 0.016183382 0.049052379 0.031573773 0.015393986 0.036242406
#>  [121] 0.046849193 0.021152486 0.049505774 0.032130589 0.026927611 0.036952542
#>  [127] 0.038815683 0.022261035 0.024626903 0.021850683 0.044395619 0.071186827
#>  [133] 0.024628620 0.132582598 0.014268788 0.030805988 0.037358309 0.075772654
#>  [139] 0.021228888 0.023552508 0.009870647 0.014938937 0.049275350 0.033451907
#>  [145] 0.024646486 0.037239571 0.022227519 0.016880463 0.019368825 0.037878222
#>  [151] 0.030322650 0.054423147 0.035293132 0.038938309 0.029035822 0.038862841
#>  [157] 0.022669201 0.049406363 0.077602123 0.023004712 0.026354699 0.018682315
#>  [163] 0.015141243 0.026154914 0.025786929 0.037677225 0.038002285 0.033654539
#>  [169] 0.039753445 0.018947211 0.021253047 0.047866695 0.067131185 0.019266302
#>  [175] 0.024409646 0.072838308 0.061120732 0.024328585 0.044723798 0.040268813
#>  [181] 0.099582360 0.036908753 0.098276980 0.043310835 0.023331342 0.040157376
#>  [187] 0.031604795 0.032224952 0.056002207 0.017713882 0.052844883 0.013659261
#>  [193] 0.019532205 0.024940335 0.064903067 0.028848184 0.028588482 0.021935059
#>  [199] 0.036286925 0.023431448 0.033076402 0.018029893 0.027007145 0.035293409
#>  [205] 0.015866822 0.043418345 0.029986357 0.027571471 0.035102485 0.013626934
#>  [211] 0.065269631 0.024321617 0.059698439 0.036343053 0.030796180 0.052037126
#>  [217] 0.043484893 0.019065970 0.018144306 0.062371898 0.031139793 0.045880213
#>  [223] 0.023748805 0.039605674 0.022528409 0.022221017 0.034407152 0.018990269
#>  [229] 0.017925586 0.049093108 0.071387221 0.044633163 0.035063331 0.030401115
#>  [235] 0.046116904 0.022594751 0.016963178 0.036390696 0.023150494 0.039717572
#>  [241] 0.036170118 0.030396748 0.016043514 0.031534265 0.025262352 0.031989039
#>  [247] 0.027053469 0.038374852 0.165369409 0.032432815 0.038367019 0.029496397
#>  [253] 0.034300183 0.039323904 0.025684435 0.064233203 0.028375632 0.023067373
#>  [259] 0.022396883 0.064022471 0.047092534 0.029160167 0.063469123 0.025585981
#>  [265] 0.018803174 0.106748727 0.015545788 0.065937214 0.029814737 0.040245590
#>  [271] 0.021175747 0.034096323 0.016550319 0.026087001 0.015323319 0.041749519
#>  [277] 0.042567561 0.027924072 0.029426402 0.015919935 0.016718572 0.078094452
#>  [283] 0.029994265 0.021575798 0.085740838 0.021742917 0.015094257 0.031262050
#>  [289] 0.017266605 0.034012817 0.040324497 0.056510927 0.036156535 0.020367333
#>  [295] 0.018189828 0.075538449 0.017298081 0.040483015 0.029049461 0.015192150
#>  [301] 0.013232978 0.047481779 0.023070467 0.025947950 0.017109682 0.033002674
#>  [307] 0.034583784 0.034377588 0.088680209 0.034313634 0.029827857 0.025317881
#>  [313] 0.032286320 0.022323735 0.031879957 0.029632664 0.013038147 0.121791252
#>  [319] 0.013190308 0.027889446 0.048937652 0.022565966 0.055955903 0.017523112
#>  [325] 0.039445367 0.034680876 0.015955323 0.022103193 0.032310938 0.033634768
#>  [331] 0.062778279 0.018181650 0.050501888 0.022150223 0.031260785 0.026685371
#>  [337] 0.043896235 0.037200259 0.025404329 0.014830317 0.018769920 0.033809397
#>  [343] 0.059614319 0.122309107 0.023608598 0.045231377 0.081291825 0.038624248
#>  [349] 0.069733879 0.016207674 0.022015636 0.047393736 0.018534416 0.018788335
#>  [355] 0.070428546 0.033644604 0.056688660 0.017745677 0.030025553 0.031018054
#>  [361] 0.020064156 0.021324832 0.029055715 0.036072891 0.072597058 0.044343967
#>  [367] 0.045949598 0.037890998 0.025002881 0.015905389 0.032757190 0.059800116
#>  [373] 0.024702989 0.030423877 0.039096335 0.023558591 0.027298135 0.027409913
#>  [379] 0.043404537 0.014976029 0.035891581 0.022768952 0.031714521 0.015142483
#>  [385] 0.052263837 0.035463483 0.023407044 0.040004416 0.092042198 0.040346162
#>  [391] 0.042689050 0.023834860 0.020291935 0.041431232 0.053536886 0.034016924
#>  [397] 0.021902626 0.048657075 0.022067331 0.095161855 0.018625504 0.038560403
#>  [403] 0.022000057 0.034122338 0.039192690 0.026654738 0.024699435 0.049710857
#>  [409] 0.017238474 0.037514203 0.024213450 0.013010076 0.074250661 0.011608602
#>  [415] 0.038361904 0.022130512 0.090908615 0.031149932 0.043097022 0.030426331
#>  [421] 0.020188767 0.018094255 0.063153242 0.015493095 0.072732872 0.037153184
#>  [427] 0.038827014 0.021826178 0.056955818 0.026234069 0.033940744 0.029471372
#>  [433] 0.020925097 0.022731165 0.031486230 0.040587097 0.037593548 0.051489415
#>  [439] 0.014475346 0.039269138 0.046442126 0.029975152 0.066453742 0.046601606
#>  [445] 0.067087264 0.077727654 0.016354918 0.033259575 0.025952227 0.024567931
#>  [451] 0.009900182 0.027090182 0.023475372 0.054217300 0.058912256 0.018959961
#>  [457] 0.026736421 0.037363459 0.059830221 0.047323226 0.049872151 0.033238541
#>  [463] 0.056749927 0.031536042 0.019003230 0.067313931 0.017606757 0.020613931
#>  [469] 0.020750540 0.021157595 0.021039495 0.013068297 0.044566203 0.046678470
#>  [475] 0.044999854 0.035079156 0.058265736 0.031672891 0.019991761 0.048735256
#>  [481] 0.022197269 0.044226862 0.047685717 0.021094507 0.048025304 0.057466139
#>  [487] 0.022445726 0.027986211 0.027177529 0.040142916 0.010315679 0.027182305
#>  [493] 0.030614948 0.019560540 0.025328522 0.035004501 0.044966834 0.053743439
#>  [499] 0.046800138 0.024302076 0.018905814 0.037290930 0.039470931 0.025086939
#>  [505] 0.032099623 0.032952111 0.025347700 0.020929041 0.039616620 0.020078392
#>  [511] 0.049459636 0.019081814 0.023858907 0.039563280 0.037318498 0.028495121
#>  [517] 0.016611969 0.022342445 0.051614123 0.041913773 0.041645630 0.041054041
#>  [523] 0.033852355 0.051421974 0.040337690 0.031902105 0.025627825 0.068488922
#>  [529] 0.018314672 0.067162614 0.028781447 0.023681286 0.123327896 0.030575730
#>  [535] 0.027758423 0.014677220 0.016651792 0.028957171 0.017285950 0.030314567
#>  [541] 0.028414973 0.024999522 0.034189825 0.061323357 0.018885406 0.028104625
#>  [547] 0.030837610 0.015831620 0.028292108 0.019190894 0.023566713 0.027972849
#>  [553] 0.016762374 0.051339880 0.026230821 0.050714530 0.022445698 0.022872428
#>  [559] 0.018733607 0.030677366 0.023628276 0.031765725 0.040514944 0.017124182
#>  [565] 0.043285177 0.040017508 0.034676741 0.035644133 0.025271135 0.035497377
#>  [571] 0.056591867 0.044004766 0.021744731 0.039289775 0.014022580 0.026433185
#>  [577] 0.017524935 0.048730023 0.022528059 0.058113883 0.061553989 0.085271092
#>  [583] 0.019442570 0.037034950 0.016222929 0.022159741 0.053548255 0.027552941
#>  [589] 0.034317964 0.067907503 0.029784410 0.020986919 0.027714258 0.044158484
#>  [595] 0.023911987 0.015159746 0.022764138 0.030927071 0.025627719 0.037440702
#>  [601] 0.042433090 0.022369239 0.026126290 0.068525805 0.018634602 0.019069490
#>  [607] 0.014239249 0.057549950 0.036695395 0.017944413 0.025718997 0.015867026
#>  [613] 0.022015570 0.097182992 0.079842502 0.026300145 0.034789798 0.042938102
#>  [619] 0.044827071 0.021084461 0.016999118 0.019962556 0.055442332 0.023878417
#>  [625] 0.015372725 0.033388976 0.035988023 0.020004310 0.069791680 0.015412416
#>  [631] 0.024342632 0.035893583 0.028254486 0.056707898 0.032714845 0.023107911
#>  [637] 0.047072379 0.030226683 0.029825055 0.073044139 0.022136957 0.033815441
#>  [643] 0.018112696 0.015126358 0.014313301 0.022697159 0.010328643 0.037704908
#>  [649] 0.018676511 0.043494720 0.085718761 0.033627605 0.032386997 0.037162822
#>  [655] 0.033566141 0.065910877 0.019747958 0.027720435 0.025536812 0.040572880
#>  [661] 0.040415912 0.030101606 0.027580307 0.031889589 0.039202397 0.025129226
#>  [667] 0.028552276 0.024181543 0.052201864 0.041756660 0.062173770 0.014914101
#>  [673] 0.087966819 0.032754233 0.022429114 0.012931334 0.027034510 0.044816295
#>  [679] 0.019664817 0.064143310 0.042911787 0.018436747 0.012354162 0.031711435
#>  [685] 0.043139841 0.039796326 0.012738805 0.013200770 0.046820381 0.047885512
#>  [691] 0.039049154 0.027115678 0.045673419 0.021920857 0.012924310 0.026526290
#>  [697] 0.024115318 0.044204543 0.017384428 0.047546605 0.033240126 0.029199149
#>  [703] 0.019267446 0.021478003 0.055451053 0.073103239 0.029510308 0.017254072
#>  [709] 0.016386792 0.039323463 0.027617563 0.024273640 0.023613457 0.051983801
#>  [715] 0.031060840 0.043045559 0.017438609 0.045342534 0.019937135 0.014043089
#>  [721] 0.029961828 0.035402079 0.021945272 0.043089017 0.019304845 0.027344105
#>  [727] 0.025273622 0.036217524 0.020154243 0.031557380 0.043665200 0.022392178
#>  [733] 0.029003892 0.026661595 0.028268724 0.017731544 0.036365228 0.037590983
#>  [739] 0.024023088 0.061040426 0.051732050 0.039960146 0.011180310 0.020919941
#>  [745] 0.050056150 0.043495353 0.062347003 0.050559182 0.034812076 0.026055840
#>  [751] 0.048767602 0.032205489 0.050427697 0.121068958 0.040662750 0.029392138
#>  [757] 0.028748396 0.043847098 0.026477978 0.025098050 0.029934287 0.045469439
#>  [763] 0.031528786 0.031962657 0.032162836 0.020804199 0.071277549 0.049837786
#>  [769] 0.016351910 0.026119375 0.027371923 0.051042931 0.034279272 0.034077870
#>  [775] 0.016312355 0.027003239 0.032467373 0.046699908 0.055411762 0.026437986
#>  [781] 0.016300214 0.023607129 0.014325850 0.062463356 0.017860217 0.018283688
#>  [787] 0.009068809 0.048604754 0.031777911 0.024295402 0.033203813 0.019555348
#>  [793] 0.015501446 0.029450749 0.021273901 0.037311286 0.010756197 0.029385567
#>  [799] 0.022814375 0.026653289 0.032768917 0.015287288 0.115500461 0.086313427
#>  [805] 0.036794079 0.106541425 0.060093532 0.052793249 0.050974161 0.020261958
#>  [811] 0.023486140 0.058347894 0.025636169 0.051855059 0.036958618 0.024454919
#>  [817] 0.037423653 0.013021005 0.011981446 0.032939145 0.017537523 0.025925862
#>  [823] 0.047715949 0.013043021 0.037685012 0.058643188 0.025155865 0.016652845
#>  [829] 0.030444499 0.016521399 0.049655394 0.027031990 0.017514216 0.038504362
#>  [835] 0.018364696 0.031857682 0.044113748 0.024698583 0.059535861 0.017151192
#>  [841] 0.025927885 0.019228355 0.023742356 0.021415883 0.019150627 0.032053295
#>  [847] 0.022644319 0.025060624 0.052711439 0.012278843 0.030120917 0.090108575
#>  [853] 0.032610475 0.040920940 0.030519326 0.020055716 0.038205578 0.087121665
#>  [859] 0.026258249 0.026671243 0.035106422 0.106156141 0.029528263 0.012763049
#>  [865] 0.029416150 0.015138383 0.054356943 0.018750546 0.025837143 0.045207247
#>  [871] 0.018936695 0.048020546 0.021523544 0.028667719 0.080926812 0.014543367
#>  [877] 0.038397201 0.052499562 0.017771142 0.032841483 0.035910937 0.038893470
#>  [883] 0.039933501 0.037382199 0.032799137 0.028350266 0.034443566 0.034746997
#>  [889] 0.024486623 0.019199368 0.054317282 0.031814078 0.027024206 0.014744766
#>  [895] 0.022248763 0.107176490 0.031007907 0.027418586 0.067775736 0.020731573
#>  [901] 0.090771979 0.048902134 0.023566653 0.027334176 0.052586061 0.031051511
#>  [907] 0.031012100 0.030268691 0.023634920 0.044365062 0.015237004 0.021735378
#>  [913] 0.022804107 0.018862152 0.039393820 0.033774362 0.037826043 0.072381188
#>  [919] 0.013886611 0.052694248 0.033176674 0.022474040 0.029782260 0.020847874
#>  [925] 0.034400714 0.017468774 0.043531103 0.009869482 0.035076948 0.040141235
#>  [931] 0.046970931 0.026035316 0.033039499 0.070418370 0.025290418 0.035672461
#>  [937] 0.017087505 0.020189796 0.016600071 0.050659587 0.009284349 0.059028910
#>  [943] 0.038293720 0.043167145 0.037369457 0.035585243 0.032413255 0.047749551
#>  [949] 0.025533334 0.026082981 0.047761920 0.032728343 0.033501342 0.038320800
#>  [955] 0.027262168 0.024716633 0.015691833 0.021635584 0.038276030 0.016248643
#>  [961] 0.019905369 0.042670872 0.015602092 0.021668876 0.016193316 0.021489905
#>  [967] 0.038470593 0.048131364 0.040989248 0.015450334 0.015976226 0.035708658
#>  [973] 0.022575560 0.031872338 0.021543428 0.019651630 0.040294385 0.033846606
#>  [979] 0.024842856 0.037966122 0.015812280 0.022763943 0.028552082 0.050232822
#>  [985] 0.028696164 0.041800152 0.076667679 0.022958579 0.020056976 0.010116412
#>  [991] 0.028451756 0.024314304 0.020667686 0.050010203 0.019810984 0.018628473
#>  [997] 0.029953792 0.064570433 0.043172516 0.041080499
#>