Compute Spacing Measuring Functions

This function is used to compute discretized functions, i.e. vectors, to explore functional spaces.

Usage

space_measuring_functions(X, M = 20, space = "BM")

Arguments

X

A dfts object or data which can be automatically converted to that format. See dfts().

M

Integer for the number of functions to generate.

space

String for the space of interest. Options are Brownian motion ('BM'), principal components ('PC'), and vectors in iid standard, random normals ('RN'). Additional options are forthcoming

Value

Data.frame with columns of discretized functions describing the space. Columns are independent functions.

See also

Examples

space_measuring_functions(M = 10, space = "BM", X = electricity)
#>               [,1]         [,2]        [,3]         [,4]        [,5]
#>  [1,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#>  [2,]  0.078001258 -0.171349465  0.12742486 -0.196639685  0.05516466
#>  [3,]  0.253053660 -0.202344557 -0.01633313 -0.149734673 -0.05002370
#>  [4,]  0.092199836 -0.105931430 -0.11420821 -0.171400717  0.06726889
#>  [5,]  0.063398747  0.136527817 -0.23801983 -0.205748501 -0.17258595
#>  [6,]  0.188821972 -0.001141092  0.03060114 -0.150644905 -0.97253284
#>  [7,] -0.162313110  0.231114112  0.06292841 -0.353230451 -0.87721631
#>  [8,] -0.089546938  0.175141302  0.14370289 -0.079470461 -0.37887023
#>  [9,]  0.261584854  0.110710609  0.04042730 -0.125683968 -0.47211747
#> [10,] -0.146686212 -0.075302614 -0.21859257 -0.358792137 -0.25725992
#> [11,] -0.083371764 -0.161113734  0.26509796 -0.432591858 -0.36740940
#> [12,]  0.157745114 -0.121278691 -0.01012542 -0.161162451 -0.43414008
#> [13,] -0.032988361 -0.243160027 -0.14245002 -0.291100096 -0.30137833
#> [14,]  0.259987047 -0.339774195 -0.24474682 -0.282577852 -0.29546694
#> [15,] -0.089765872 -0.158071627 -0.48475143 -0.093378767 -0.41976945
#> [16,] -0.246591800 -0.257519775 -0.41828340  0.020455691 -0.34892386
#> [17,]  0.008615776 -0.147166420 -0.60993303  0.006237936 -0.21744057
#> [18,]  0.047790860 -0.004543640 -0.30239849  0.137984804  0.11176375
#> [19,] -0.040011134  0.028451383 -0.29294721  0.225642528  0.04630933
#> [20,] -0.047250990 -0.101190703 -0.12444760  0.323067722  0.18831404
#> [21,] -0.039875365 -0.066313572 -0.35076761  0.276543166  0.39179475
#> [22,] -0.021683592  0.102087806 -0.29352052  0.464202942  0.44182851
#> [23,] -0.161991322  0.339324773 -0.17689579  0.806367766  0.33016185
#> [24,] -0.253144200 -0.037932890  0.11645772  0.503683351  0.24927709
#>              [,6]         [,7]        [,8]         [,9]        [,10]
#>  [1,]  0.00000000  0.000000000  0.00000000  0.000000000  0.000000000
#>  [2,]  0.13842097  0.003430832 -0.15943614 -0.014860070 -0.058858336
#>  [3,]  0.30238813 -0.292610329 -0.29296958 -0.335369213 -0.146564039
#>  [4,]  0.20836161  0.088738971 -0.45133904 -0.333876299 -0.134823327
#>  [5,] -0.02907600  0.077200869 -0.55538430 -0.414993593 -0.151280179
#>  [6,] -0.18514063  0.016351716 -0.48238686 -0.157231200 -0.014713781
#>  [7,]  0.02605086  0.012739571 -0.71945668 -0.047946965  0.225704497
#>  [8,]  0.37405285  0.140528061 -0.96189948 -0.066990228  0.008375134
#>  [9,]  0.25600140  0.087387995 -1.01781593  0.162130175 -0.272125951
#> [10,]  0.53351589 -0.222758043 -0.95543007  0.059727156 -0.376090354
#> [11,]  0.21162480 -0.246788989 -0.74028042  0.041213865 -0.220799537
#> [12,]  0.39496637 -0.040848938 -0.49444418  0.088763546 -0.089773938
#> [13,]  0.36724032  0.039932832 -0.56192026  0.123911408 -0.211390305
#> [14,]  0.39714337  0.290339771 -0.57345590  0.298551225 -0.312139998
#> [15,]  0.44345087  0.224001609 -0.49009371 -0.007347947 -0.543529982
#> [16,]  0.47412043  0.388355451 -0.46767348  0.016875160 -0.521768568
#> [17,]  0.37376405  0.371740130 -0.47992230  0.117218187 -0.578619069
#> [18,]  0.10787273  0.056320443 -0.17069044 -0.108333855 -0.664595980
#> [19,] -0.20065886  0.418236309 -0.04785981  0.365829683 -0.649893395
#> [20,] -0.15969953  0.548672014 -0.58772902  0.232289276 -0.593650837
#> [21,] -0.39602285  0.627075419 -1.01807661  0.245306270 -0.532503829
#> [22,]  0.26507032  0.604072546 -0.86008258  0.283946801 -0.312910458
#> [23,]  0.39362130  0.656326862 -0.91534660  0.324382200 -0.427845622
#> [24,]  0.57893608  0.526250093 -1.05487117  0.154622278 -0.533527105
space_measuring_functions(M = 10, space = "PC", X = electricity)
#>             V1           V2          V3          V4          V5         V6
#> 1    146.68738    -8.147271  -857.04082   -58.46399   -99.75877  -356.5924
#> 2    361.84495   128.014796   706.64380 -1024.69043  1476.59352  -975.4319
#> 3    927.91943 -2054.157073   142.78603   477.83078 -1142.35369  -480.4093
#> 4   1320.11248  -164.493674 -1141.22770  -318.39312   689.62647 -1176.8475
#> 5    559.12121  -149.885713  -718.39386    72.14414   248.23730  -692.0707
#> 6    998.68341   -55.893519 -1501.58501  -115.19965 -1239.16444 -2341.4307
#> 7    158.04052  -365.914914   480.34949  1123.23523   -64.00014   159.6705
#> 8    -25.23838  -517.545103   763.39323 -1312.69805  -593.84958   905.9648
#> 9    138.50518 -1601.746852   140.45119  1981.38529  1230.03973  1143.0585
#> 10 -1102.89599    -4.355295   -47.37350  1071.42055   -13.09542 -1034.3718
#> 11 -1739.81857   372.539832   830.31611  -502.49262  -339.82662  -143.0574
#> 12  1773.31196  1638.208496   443.64036   560.26313  -170.52545 -1549.8707
#> 13 -1168.20483  1506.774640    16.08534  -166.85966   476.25809   613.4910
#> 14   245.61424  -282.140860   226.16997  -632.69545  -188.69584  1056.5878
#> 15   802.41649  1014.918868   864.78984  -606.97973   646.37779 -1792.0119
#> 16  -779.46109  -926.404413   700.07770 -1002.67672  1068.91701  -543.4961
#> 17  -175.50579   274.386056 -1363.19452 -1201.14368  1209.46229   458.7319
#> 18  -304.03318 -1028.257379    71.29990  1151.05035  -394.76878 -1057.0003
#> 19   130.66722   375.367930  -354.82867 -1184.69822   659.33384  -117.5622
#> 20 -1084.61348  -741.461674  -828.98513  -203.46038  -956.92022   328.5583
#> 21  1106.54395  1262.928876  -603.42127   390.60913  1760.99046 -1521.9477
#> 22   809.62453   403.644559   257.11091   716.00084  1077.61423  -794.5305
#> 23  -434.79316  -556.945400  1352.66753 -1118.20050   611.85995  -330.4368
#> 24  1021.71302  -712.559479  -306.78116  -132.13000   667.38079   334.2143
#>            V7         V8           V9        V10
#> 1   1335.6109   315.0484   448.097605   840.2469
#> 2    325.9436    27.8203  -307.795939   435.8517
#> 3  -1063.1439   921.0491   433.119859   324.8692
#> 4    972.9142  -704.2226  -367.289297 -1380.0860
#> 5   -870.1990  1265.4994  -493.221185  -219.9842
#> 6  -1354.6273   838.0989    -7.507341  -183.7419
#> 7    170.1528   199.4609  -821.843013   -26.7916
#> 8    315.8206   539.1646   573.038312  -942.8285
#> 9   1355.1138  -358.1488  -300.714855  -688.4177
#> 10   107.7776  -600.3925   574.378396  1636.3607
#> 11  -163.4082   174.2887  -663.137789  -320.7157
#> 12  -752.2276  -571.6366   275.167355  -669.2234
#> 13   861.4064 -1154.1089  1782.872765  -610.1286
#> 14  2257.7097 -1386.9069   602.148774  -631.2297
#> 15   723.5664  1569.6106   741.423495 -2082.0083
#> 16   -71.4953   367.8151  1021.392478   313.3734
#> 17 -1009.8972  -728.1749  -108.119635   225.8127
#> 18 -1496.6838 -2918.8029  -170.542391  -112.5641
#> 19 -2594.1546  -800.1695  1435.613103   613.8603
#> 20   293.0887 -2020.4627   356.221249   563.5511
#> 21  -338.8389   573.7229  -570.108960  -357.2595
#> 22  -117.5313   125.4535 -1969.038525 -1788.7135
#> 23  -547.7483  -125.0167   -75.938427   418.5502
#> 24   729.5815   709.1740   398.505815  -655.5780