Compute Spacing Measuring Functions

This function is used to compute discretized functions, i.e. vectors, to explore functional spaces.

Usage

space_measuring_functions(X, M = 20, space = "BM")

Arguments

X

A dfts object or data which can be automatically converted to that format. See dfts().

M

Integer for the number of functions to generate.

space

String for the space of interest. Options are Brownian motion ('BM'), principal components ('PC'), and vectors in iid standard, random normals ('RN'). Additional options are forthcoming

Value

Data.frame with columns of discretized functions describing the space. Columns are independent functions.

See also

Examples

space_measuring_functions(M=10, space="BM", X=electricity)
#>             [,1]        [,2]       [,3]         [,4]        [,5]         [,6]
#>  [1,] 0.00000000  0.00000000 0.00000000  0.000000000  0.00000000  0.000000000
#>  [2,] 0.04483331 -0.23531378 0.01748255 -0.005287588 -0.13454487  0.126546511
#>  [3,] 0.11595179 -0.22547381 0.38820990  0.234233383 -0.41717508  0.439007918
#>  [4,] 0.38969431 -0.08525194 0.31029867  0.248285557 -0.66568707  0.307700714
#>  [5,] 0.60476414  0.09706558 0.49775857  0.477838969 -0.65994288  0.035460230
#>  [6,] 0.97624447  0.07519261 0.27679031  0.504379134 -0.80556518  0.075283731
#>  [7,] 0.95636050  0.16867714 0.32504244  0.605156518 -0.84081854 -0.213080588
#>  [8,] 1.22174223  0.38906866 0.12522662  0.429154286 -0.62019598 -0.061851002
#>  [9,] 1.06959153  0.18989154 0.17568618  0.778635680 -0.45959814  0.005662566
#> [10,] 1.11734385  0.08315524 0.33023460  0.621807316 -0.45392853  0.285925674
#> [11,] 0.91368824  0.32452950 0.51754087  0.397041525 -0.29496201  0.233639908
#> [12,] 0.53793680  0.07641312 0.32296826  0.206484886 -0.13239751  0.255722237
#> [13,] 0.65801377 -0.15157853 0.36516521  0.269937099 -0.01840644  0.337452654
#> [14,] 0.57196672  0.38298398 0.77977946  0.344539104 -0.29894345  0.333151937
#> [15,] 0.51469892  0.33275098 0.68789452  0.571112474 -0.12632360  0.069186307
#> [16,] 0.52707017  0.41476519 0.48930610  0.141988845 -0.47083284  0.183583702
#> [17,] 0.37026309  0.49892930 0.34060191 -0.143695863 -0.36204089  0.246852783
#> [18,] 0.35649584  0.48015349 0.49997145 -0.157048839 -0.76211027  0.099800836
#> [19,] 0.93737727  0.64251757 0.93050758 -0.009406726 -0.70853023 -0.054749440
#> [20,] 1.21217663  0.72945291 1.04516950  0.118753343 -0.70665824 -0.296777677
#> [21,] 1.11496291  1.05123321 0.66815926 -0.013084037 -0.63661740 -0.521752893
#> [22,] 1.24997971  1.37735938 0.58592840 -0.164040152 -0.59312672 -0.643703121
#> [23,] 1.42503105  1.83799424 0.56876657 -0.130585591 -0.93354523 -0.617698477
#> [24,] 1.28892454  2.08903635 0.94177790 -0.005286767 -0.88062575 -0.912295011
#>              [,7]         [,8]       [,9]        [,10]
#>  [1,]  0.00000000  0.000000000  0.0000000  0.000000000
#>  [2,]  0.13896155  0.189800784 -0.2924194  0.308204593
#>  [3,]  0.17583498 -0.091331842 -0.4825476  0.001262871
#>  [4,]  0.03963221 -0.172371955 -0.4300743 -0.341356531
#>  [5,] -0.25504458  0.286716577 -0.6137996 -0.095625604
#>  [6,] -0.33599178  0.001913033 -0.5179507 -0.248867211
#>  [7,] -0.35034261  0.002132354 -0.2189594  0.125439887
#>  [8,] -0.27789613  0.018349882 -0.1696612  0.225118935
#>  [9,] -0.36050088 -0.065167557  0.1876386  0.252759397
#> [10,]  0.08268281 -0.406533891  0.2683451  0.229063960
#> [11,]  0.29686288 -0.356526333  0.1473073 -0.201318139
#> [12,]  0.27403130 -0.337392652  0.2200993  0.079823741
#> [13,]  0.05798018 -0.457002046  0.6495148  0.088887203
#> [14,]  0.06628319 -0.562233591  0.3909807  0.119050353
#> [15,]  0.43755550 -0.400419278  0.2211688 -0.188391741
#> [16,]  0.55494394 -0.167223616  0.1333655 -0.467463963
#> [17,]  0.73097995  0.154028398 -0.3279489 -0.726562129
#> [18,]  0.29113533  0.390736646 -0.6368692 -0.525240395
#> [19,]  0.47964217  0.456632978 -0.5611293 -0.284930114
#> [20,]  0.74401008  0.150666270 -0.6667592 -0.055370215
#> [21,]  0.57796975  0.409023199 -0.5869687  0.118683681
#> [22,]  0.70593894  0.392261357 -0.4250564  0.231404651
#> [23,]  0.62928796  0.427266368 -0.8291719  0.513180799
#> [24,]  0.86133499  0.272399443 -0.7550032  0.233572251
space_measuring_functions(M=10, space="PC", X=electricity)
#>             V1          V2          V3         V4          V5          V6
#> 1   1841.68149  -469.27809  -924.88788  506.27375  -112.42018  -543.75228
#> 2  -1518.98989  -376.43435  1131.32132 -405.16453  -236.61000  -371.72555
#> 3    169.05668   -19.34726   595.09734 -237.32796  1642.78620   469.82390
#> 4    294.86113  -334.34670  -188.04075 -273.07138  -966.10721  -790.70527
#> 5    390.44710  -296.58630   202.46920 -208.31527  -129.11465  -153.88214
#> 6   -881.64055 -1514.77234   395.53474  194.27765  -284.14766   -15.61158
#> 7    704.31389  1435.36309   695.14793 1071.79057  -225.02922  1153.97647
#> 8    694.95111   865.10988   776.20042 -460.68751  -449.77357  1061.42063
#> 9   -540.85112   744.44335   364.65673 -104.85439   614.92098   156.92543
#> 10   822.35150   269.04780   251.01271  513.81186 -1102.20832   415.27777
#> 11 -1374.57054   288.49605 -1011.80369   82.43859 -1362.06803 -1309.25709
#> 12   223.34283  1010.04233  -284.27635 -742.97533 -1290.34324 -1454.91864
#> 13  -691.90966  1214.19978   123.26497 -889.55804   -32.95747  1526.55036
#> 14   868.65861 -1006.60116  1036.56323  665.21433 -2740.81616  -749.08651
#> 15  -978.02288  1551.93290  1628.16497  865.07194   486.16926 -1462.47676
#> 16 -1241.68900    74.14735   364.61768 2208.53886  1379.03952  -181.08817
#> 17   153.67647    49.34635  -724.35578 -475.11845  2165.84281   856.69424
#> 18    76.05319  -302.93970  2807.63925  252.97060   414.12001  1642.48479
#> 19   277.25797   -72.60567  1355.05030 -760.11749   344.56572  -569.81738
#> 20  1025.98732  -198.71377 -1034.51089 -220.70240  -515.01590 -1386.57214
#> 21  1174.02029   -40.75791  -317.92066 1024.32744   656.06255   796.22871
#> 22   178.07986   989.79613   -30.94968  605.99213   715.78898  -465.05927
#> 23  1096.01040   428.87824 -1211.13264 -659.08525 -1414.14538  -227.42251
#> 24   529.03850   727.68101   262.17483 -120.86629   854.07288   308.88308
#>             V7           V8           V9          V10
#> 1    163.81307  1707.975846 -1075.576683  1149.419926
#> 2    255.03554   665.705591    35.400462 -1477.041940
#> 3    -99.05553   877.864295   -83.421489   764.368672
#> 4   -224.27192  -686.391634  -666.126732  -981.195029
#> 5   -443.31656  1137.773875   818.311105   721.437685
#> 6  -1238.68549  -570.476610  -215.855358    -7.821575
#> 7    109.30335  -141.529478    -4.015365   389.113087
#> 8    518.26571     7.437739  -412.981790   816.092428
#> 9   -466.84344  -894.875291  -879.535198  -650.805318
#> 10  -261.65740   453.986971   291.642341  -314.708023
#> 11   -79.20859   581.081616  -702.930157  1130.991419
#> 12   161.91903   923.009962  1287.762645  -634.526876
#> 13  -693.99018 -1553.146018  1155.600888  -702.795351
#> 14  -325.21715  -713.522471  -304.526192 -1520.823676
#> 15   729.90662 -1489.162642  -187.706135  -420.944360
#> 16  -477.13893   821.939016   106.792501   136.221645
#> 17    12.57217   803.442009 -2190.022581 -1517.811666
#> 18  -169.81993   -48.185784  -385.905520   342.454624
#> 19  -317.19058  1028.805531  -807.325491    -7.662578
#> 20  -842.90338   411.700419   602.037243  2087.316964
#> 21   528.86699   223.716019   290.313568  -875.620161
#> 22  -990.70648  1345.146303   -83.885219   228.409813
#> 23  -939.31786   148.001931   750.582723  -334.689901
#> 24  1007.82806    29.225712 -1024.512274  -875.271573